+++ README.md
@@ -1,3 +1,7 @@
+npm run init-db
+npm run seed-db
+
+
 This is a [Next.js](https://nextjs.org) project bootstrapped with [`create-next-app`](https://nextjs.org/docs/app/api-reference/cli/create-next-app).
 
 ## Getting Started
