+++ package.json
@@ -6,12 +6,16 @@
     "dev": "next dev --turbopack",
     "build": "next build",
     "start": "next start",
-    "lint": "next lint"
+    "lint": "next lint",
+    "init-db": "ts-node --project tsconfig.scripts.json scripts/init-db.ts",
+    "seed-db": "ts-node --project tsconfig.scripts.json scripts/seed-db.ts"
   },
   "dependencies": {
     "@radix-ui/react-select": "^2.1.6",
+    "dotenv": "^16.4.7",
     "lucide-react": "^0.477.0",
     "next": "15.2.1",
+    "pg": "^8.14.1",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "react-icons": "^5.5.0",
@@ -21,11 +25,13 @@
     "@eslint/eslintrc": "^3",
     "@tailwindcss/postcss": "^4",
     "@types/node": "^20",
+    "@types/pg": "^8.11.11",
     "@types/react": "^19",
     "@types/react-dom": "^19",
     "eslint": "^9",
     "eslint-config-next": "15.2.1",
     "tailwindcss": "^4",
+    "ts-node": "^10.9.2",
     "typescript": "^5"
   }
 }
