+++ src/core/profile/profileService.ts
@@ -1,2842 +1,49 @@
-//src/app/services/profileService.ts
-// 仮のサービス層実装（実際のDB接続などは実装によって異なります）
-
-import {
-  StudentData,
-  Friend,
-  TestData,
-} from "@/app/(pages)/profile/[subject]/page";
-import {
-  CardData,
-  ProfileData,
-} from "@/app/(pages)/profile/_components/TestResultCard";
-
-// 1000: ⚪︎
-// 1001: ×
-// 1002: 解答する必要が無い問題
-
-// モックデータ
-const mockStudentData_R_2018: StudentData = {
-  name: "あなた",
-  score: 150,
-  percentage: 75,
-  targetPercentage: 70,
-  date: "2024/01/26",
-  memo: "今回は実力が出せなかっただけなんだ。本当の実力はこんなもんじゃないよ。",
-  sectionTotals: {
-    1: 12, // 大問1の小計
-    2: 47, // 大問2の小計
-    3: 17, // 大問3の小計
-    4: 35, // 大問4の小計
-    5: 18, // 大問5の小計
-    6: 30, // 大問6の小計
-  },
-  sectionPercentages: {
-    1: 86,
-    2: 100,
-    3: 74,
-    4: 88,
-    5: 60,
-    6: 83,
-  },
-  targetSectionTotals: {
-    1: 12, // 大問1の小計目標
-    2: 30, // 大問2の小計目標
-    3: 18, // 大問3の小計目標
-    4: 25, // 大問4の小計目標
-    5: 18, // 大問5の小計目標
-    6: 25, // 大問6の小計目標
-  },
-  targetSectionPercentages: {
-    1: 86,
-    2: 64,
-    3: 78,
-    4: 63,
-    5: 60,
-    6: 69,
-  },
-  answers: {
-    1: 4,
-    2: 3,
-    3: 2,
-    4: 1,
-    5: 3,
-    6: 3,
-    7: 2,
-    8: 3,
-    9: 2,
-    10: 1,
-    11: 4,
-    12: 1,
-    13: 3,
-    14: 1,
-    15: 3,
-    16: 3,
-    17: 3,
-    18: 3,
-    19: 2,
-    20: 4,
-    21: 2,
-    22: 3,
-    23: 2,
-    24: 2,
-    25: 4,
-    26: 8,
-    27: 2,
-    30: 1,
-    31: 2,
-    32: 4,
-    33: 2,
-    34: 3,
-    35: 1,
-    36: 4,
-    37: 4,
-    38: 2,
-    39: 3,
-    40: 4,
-    41: 3,
-    42: 2,
-    43: 2,
-    44: 2,
-    45: 4,
-    46: 4,
-    47: 2,
-    48: 1,
-    49: 1,
-    50: 2,
-    51: 4,
-    52: 3,
-    53: 4,
-    54: 1,
-  },
-};
-
-const mockStudentData_1A_2015: StudentData = {
-  name: "あなた",
-  score: 86,
-  percentage: 86,
-  targetPercentage: 75,
-  date: "2024/01/26",
-  memo: "今回の数1Aは実力が出せなかっただけなんだ。本当の実力はこんなもんじゃないよ。",
-  sectionTotals: {
-    1: 17, // 大問1の小計
-    2: 22, // 大問2の小計
-    3: 13, // 大問3の小計
-    4: 17, // 大問4の小計
-    5: 0, // 大問5の小計
-    6: 17, // 大問6の小計
-  },
-  sectionPercentages: {
-    1: 85,
-    2: 88,
-    3: 87,
-    4: 85,
-    5: 0,
-    6: 85,
-  },
-  targetSectionTotals: {
-    1: 15, // 大問1の小計目標
-    2: 20, // 大問2の小計目標
-    3: 10, // 大問3の小計目標
-    4: 15, // 大問4の小計目標
-    5: 0, // 大問5の小計目標
-    6: 15, // 大問6の小計目標
-  },
-  targetSectionPercentages: {
-    1: 75,
-    2: 80,
-    3: 67,
-    4: 75,
-    5: 0,
-    6: 75,
-  },
-  answers: {
-    1: 1000,
-    2: 1000,
-    3: 1000,
-    4: 1000,
-    5: 1001,
-    6: 1000,
-    7: 1000,
-    8: 1000,
-    9: 1000,
-    10: 1000,
-    11: 1000,
-    12: 1000,
-    13: 1001,
-    14: 1000,
-    15: 1000,
-    16: 1000,
-    17: 1001,
-    18: 1000,
-    19: 1000,
-    20: 1000,
-    21: 1000,
-    22: 1000,
-    23: 1000,
-    24: 1000,
-    25: 1001,
-    26: 1002,
-    27: 1002,
-    28: 1002,
-    29: 1002,
-    30: 1002,
-    31: 1002,
-    32: 1002,
-    33: 1000,
-    34: 1000,
-    35: 1000,
-    36: 1000,
-    37: 1000,
-    38: 1001,
-  },
-};
-
-const mockFriendsData_R_2018: Friend[] = [
-  {
-    id: 1,
-    name: "大貫しもん",
-    score: 144,
-    percentage: 72,
-    targetPercentage: 80,
-    date: "2024/01/26",
-    memo: "今回は体調がすぐれなかった。",
-    sectionTotals: {
-      1: 12, // 大問1の小計
-      2: 38, // 大問2の小計
-      3: 17, // 大問3の小計
-      4: 35, // 大問4の小計
-      5: 18, // 大問5の小計
-      6: 30, // 大問6の小計
-    },
-    sectionPercentages: {
-      1: 86,
-      2: 81,
-      3: 74,
-      4: 88,
-      5: 60,
-      6: 83,
-    },
-    targetSectionTotals: {
-      1: 12, // 大問1の小計目標
-      2: 30, // 大問2の小計目標
-      3: 18, // 大問3の小計目標
-      4: 25, // 大問4の小計目標
-      5: 18, // 大問5の小計目標
-      6: 25, // 大問6の小計目標
-    },
-    targetSectionPercentages: {
-      1: 86,
-      2: 64,
-      3: 78,
-      4: 63,
-      5: 60,
-      6: 69,
-    },
-    answers: {
-      1: 4,
-      2: 3,
-      3: 2,
-      4: 1,
-      5: 4,
-      6: 3,
-      7: 1,
-      8: 3,
-      9: 2,
-      10: 1,
-      11: 4,
-      12: 1,
-      13: 3,
-      14: 1,
-      15: 3,
-      16: 3,
-      17: 3,
-      18: 3,
-      19: 2,
-      20: 4,
-      21: 2,
-      22: 3,
-      23: 1,
-      24: 2,
-      25: 2,
-      26: 8,
-      27: 2,
-      30: 1,
-      31: 2,
-      32: 4,
-      33: 2,
-      34: 2,
-      35: 2,
-      36: 4,
-      37: 4,
-      38: 2,
-      39: 3,
-      40: 4,
-      41: 3,
-      42: 2,
-      43: 1,
-      44: 2,
-      45: 3,
-      46: 4,
-      47: 2,
-      48: 1,
-      49: 1,
-      50: 2,
-      51: 4,
-      52: 3,
-      53: 3,
-      54: 1,
-    },
-  },
-  {
-    id: 2,
-    name: "磯辺恵美理",
-    score: 200,
-    percentage: 100,
-    targetPercentage: 100,
-    date: "2024/01/27",
-    memo: "満点取れて安心した。",
-    sectionTotals: {
-      1: 14, // 大問1の小計
-      2: 47, // 大問2の小計
-      3: 23, // 大問3の小計
-      4: 40, // 大問4の小計
-      5: 30, // 大問5の小計
-      6: 36, // 大問6の小計
-    },
-    sectionPercentages: {
-      1: 100,
-      2: 100,
-      3: 100,
-      4: 100,
-      5: 100,
-      6: 100,
-    },
-    targetSectionTotals: {
-      1: 14, // 大問1の小計目標
-      2: 47, // 大問2の小計目標
-      3: 23, // 大問3の小計目標
-      4: 40, // 大問4の小計目標
-      5: 30, // 大問5の小計目標
-      6: 36, // 大問6の小計目標
-    },
-    targetSectionPercentages: {
-      1: 100,
-      2: 100,
-      3: 100,
-      4: 100,
-      5: 100,
-      6: 100,
-    },
-    answers: {
-      1: 4,
-      2: 3,
-      3: 2,
-      4: 1,
-      5: 4,
-      6: 3,
-      7: 2,
-      8: 3,
-      9: 2,
-      10: 1,
-      11: 4,
-      12: 1,
-      13: 3,
-      14: 1,
-      15: 3,
-      16: 3,
-      17: 3,
-      18: 3,
-      19: 2,
-      20: 4,
-      21: 2,
-      22: 3,
-      23: 2,
-      24: 2,
-      25: 4,
-      26: 8,
-      27: 2,
-      30: 1,
-      31: 3,
-      32: 4,
-      33: 2,
-      34: 3,
-      35: 2,
-      36: 4,
-      37: 4,
-      38: 2,
-      39: 3,
-      40: 4,
-      41: 3,
-      42: 2,
-      43: 2,
-      44: 1,
-      45: 3,
-      46: 4,
-      47: 2,
-      48: 1,
-      49: 1,
-      50: 2,
-      51: 4,
-      52: 2,
-      53: 3,
-      54: 1,
-    },
-  },
-  {
-    id: 3,
-    name: "藤岡優仁",
-    score: 156,
-    percentage: 78,
-    targetPercentage: 90,
-    date: "2024/01/28",
-    memo: "英語長文ポラリス1を音読した成果が出てよかった。",
-    sectionTotals: {
-      1: 12, // 大問1の小計
-      2: 47, // 大問2の小計
-      3: 17, // 大問3の小計
-      4: 40, // 大問4の小計
-      5: 24, // 大問5の小計
-      6: 30, // 大問6の小計
-    },
-    sectionPercentages: {
-      1: 86,
-      2: 100,
-      3: 74,
-      4: 100,
-      5: 80,
-      6: 83,
-    },
-    targetSectionTotals: {
-      1: 12, // 大問1の小計目標
-      2: 30, // 大問2の小計目標
-      3: 18, // 大問3の小計目標
-      4: 25, // 大問4の小計目標
-      5: 18, // 大問5の小計目標
-      6: 25, // 大問6の小計目標
-    },
-    targetSectionPercentages: {
-      1: 86,
-      2: 64,
-      3: 78,
-      4: 63,
-      5: 60,
-      6: 69,
-    },
-    answers: {
-      1: 4,
-      2: 3,
-      3: 2,
-      4: 1,
-      5: 3,
-      6: 3,
-      7: 2,
-      8: 3,
-      9: 2,
-      10: 1,
-      11: 4,
-      12: 1,
-      13: 3,
-      14: 1,
-      15: 3,
-      16: 3,
-      17: 3,
-      18: 3,
-      19: 2,
-      20: 4,
-      21: 2,
-      22: 3,
-      23: 2,
-      24: 2,
-      25: 4,
-      26: 8,
-      27: 2,
-      30: 1,
-      31: 2,
-      32: 4,
-      33: 2,
-      34: 3,
-      35: 2,
-      36: 4,
-      37: 4,
-      38: 2,
-      39: 3,
-      40: 4,
-      41: 3,
-      42: 2,
-      43: 2,
-      44: 2,
-      45: 3,
-      46: 4,
-      47: 2,
-      48: 1,
-      49: 1,
-      50: 2,
-      51: 4,
-      52: 3,
-      53: 4,
-      54: 1,
-    },
-  },
-];
-
-const mockFriendsData_1A_2015: Friend[] = [
-  {
-    id: 1,
-    name: "大貫しもん",
-    score: 86,
-    percentage: 86,
-    targetPercentage: 75,
-    date: "2024/01/26",
-    memo: "今回の数1Aは体調が悪くて大変だったけど目標点を越えられてよかった。",
-    sectionTotals: {
-      1: 17, // 大問1の小計
-      2: 22, // 大問2の小計
-      3: 13, // 大問3の小計
-      4: 17, // 大問4の小計
-      5: 0, // 大問5の小計
-      6: 17, // 大問6の小計
-    },
-    sectionPercentages: {
-      1: 85,
-      2: 88,
-      3: 87,
-      4: 85,
-      5: 0,
-      6: 85,
-    },
-    targetSectionTotals: {
-      1: 15, // 大問1の小計目標
-      2: 20, // 大問2の小計目標
-      3: 10, // 大問3の小計目標
-      4: 15, // 大問4の小計目標
-      5: 0, // 大問5の小計目標
-      6: 15, // 大問6の小計目標
-    },
-    targetSectionPercentages: {
-      1: 75,
-      2: 80,
-      3: 67,
-      4: 75,
-      5: 0,
-      6: 75,
-    },
-    answers: {
-      1: 1000,
-      2: 1000,
-      3: 1000,
-      4: 1000,
-      5: 1001,
-      6: 1000,
-      7: 1000,
-      8: 1000,
-      9: 1000,
-      10: 1000,
-      11: 1000,
-      12: 1000,
-      13: 1001,
-      14: 1000,
-      15: 1000,
-      16: 1000,
-      17: 1001,
-      18: 1000,
-      19: 1000,
-      20: 1000,
-      21: 1000,
-      22: 1000,
-      23: 1000,
-      24: 1000,
-      25: 1001,
-      26: 1002,
-      27: 1002,
-      28: 1002,
-      29: 1002,
-      30: 1002,
-      31: 1002,
-      32: 1002,
-      33: 1000,
-      34: 1000,
-      35: 1000,
-      36: 1000,
-      37: 1000,
-      38: 1001,
-    },
-  },
-  {
-    id: 2,
-    name: "磯辺恵美理",
-    score: 86,
-    percentage: 86,
-    targetPercentage: 75,
-    date: "2024/01/26",
-    memo: "数1Aは死ぬと思ってたけど思ったよりも点が取れて嬉しかった。",
-    sectionTotals: {
-      1: 17, // 大問1の小計
-      2: 22, // 大問2の小計
-      3: 13, // 大問3の小計
-      4: 17, // 大問4の小計
-      5: 0, // 大問5の小計
-      6: 17, // 大問6の小計
-    },
-    sectionPercentages: {
-      1: 85,
-      2: 88,
-      3: 87,
-      4: 85,
-      5: 0,
-      6: 85,
-    },
-    targetSectionTotals: {
-      1: 15, // 大問1の小計目標
-      2: 20, // 大問2の小計目標
-      3: 10, // 大問3の小計目標
-      4: 15, // 大問4の小計目標
-      5: 0, // 大問5の小計目標
-      6: 15, // 大問6の小計目標
-    },
-    targetSectionPercentages: {
-      1: 75,
-      2: 80,
-      3: 67,
-      4: 75,
-      5: 0,
-      6: 75,
-    },
-    answers: {
-      1: 1000,
-      2: 1000,
-      3: 1000,
-      4: 1000,
-      5: 1001,
-      6: 1000,
-      7: 1000,
-      8: 1000,
-      9: 1000,
-      10: 1000,
-      11: 1000,
-      12: 1000,
-      13: 1001,
-      14: 1000,
-      15: 1000,
-      16: 1000,
-      17: 1001,
-      18: 1000,
-      19: 1000,
-      20: 1000,
-      21: 1000,
-      22: 1000,
-      23: 1000,
-      24: 1000,
-      25: 1001,
-      26: 1002,
-      27: 1002,
-      28: 1002,
-      29: 1002,
-      30: 1002,
-      31: 1002,
-      32: 1002,
-      33: 1000,
-      34: 1000,
-      35: 1000,
-      36: 1000,
-      37: 1000,
-      38: 1001,
-    },
-  },
-  {
-    id: 3,
-    name: "藤岡優仁",
-    score: 86,
-    percentage: 86,
-    targetPercentage: 75,
-    date: "2024/01/26",
-    memo: "数1Aは基礎問2周しただけだけど結構点数取れてよかった。",
-    sectionTotals: {
-      1: 15, // 大問1の小計
-      2: 19, // 大問2の小計
-      3: 7, // 大問3の小計
-      4: 15, // 大問4の小計
-      5: 0, // 大問5の小計
-      6: 13, // 大問6の小計
-    },
-    sectionPercentages: {
-      1: 75,
-      2: 76,
-      3: 54,
-      4: 75,
-      5: 0,
-      6: 65,
-    },
-    targetSectionTotals: {
-      1: 15, // 大問1の小計目標
-      2: 20, // 大問2の小計目標
-      3: 10, // 大問3の小計目標
-      4: 15, // 大問4の小計目標
-      5: 0, // 大問5の小計目標
-      6: 15, // 大問6の小計目標
-    },
-    targetSectionPercentages: {
-      1: 75,
-      2: 80,
-      3: 67,
-      4: 75,
-      5: 0,
-      6: 75,
-    },
-    answers: {
-      1: 1000,
-      2: 1000,
-      3: 1000,
-      4: 1001,
-      5: 1001,
-      6: 1000,
-      7: 1000,
-      8: 1000,
-      9: 1000,
-      10: 1000,
-      11: 1000,
-      12: 1001,
-      13: 1001,
-      14: 1000,
-      15: 1000,
-      16: 1001,
-      17: 1001,
-      18: 1000,
-      19: 1000,
-      20: 1000,
-      21: 1000,
-      22: 1000,
-      23: 1000,
-      24: 1001,
-      25: 1001,
-      26: 1002,
-      27: 1002,
-      28: 1002,
-      29: 1002,
-      30: 1002,
-      31: 1002,
-      32: 1002,
-      33: 1000,
-      34: 1000,
-      35: 1000,
-      36: 1000,
-      37: 1001,
-      38: 1001,
-    },
-  },
-];
-
-const mockTestData_R_2015: TestData = {
-  subject: "Reading",
-  year: 2015,
-  maxScore: 200,
-  testStructure: [
-    {
-      section: 1,
-      questions: [
-        { questionNumber: 1, score: 2, correctAnswer: 2 },
-        { questionNumber: 2, score: 2, correctAnswer: 1 },
-        { questionNumber: 3, score: 2, correctAnswer: 1 },
-        { questionNumber: 4, score: 2, correctAnswer: 2 },
-        { questionNumber: 5, score: 2, correctAnswer: 4 },
-        { questionNumber: 6, score: 2, correctAnswer: 4 },
-        { questionNumber: 7, score: 2, correctAnswer: 3 },
-      ],
-      sectionTotal: { score: 14 },
-    },
-    {
-      section: 2,
-      questions: [
-        { questionNumber: 8, score: 2, correctAnswer: 1 },
-        { questionNumber: 9, score: 2, correctAnswer: 3 },
-        { questionNumber: 10, score: 2, correctAnswer: 3 },
-        { questionNumber: 11, score: 2, correctAnswer: 2 },
-        { questionNumber: 12, score: 2, correctAnswer: 1 },
-        { questionNumber: 13, score: 2, correctAnswer: 3 },
-        { questionNumber: 14, score: 2, correctAnswer: 1 },
-        { questionNumber: 15, score: 2, correctAnswer: 4 },
-        { questionNumber: 16, score: 2, correctAnswer: 2 },
-        { questionNumber: 17, score: 2, correctAnswer: 4 },
-        { questionNumber: 18, score: 4, correctAnswer: 5 },
-        { questionNumber: 19, score: 4, correctAnswer: 6 },
-        { questionNumber: 20, score: 4, correctAnswer: 5 },
-        { questionNumber: 21, score: 4, correctAnswer: 1 },
-        { questionNumber: 22, score: 4, correctAnswer: 4 },
-        { questionNumber: 23, score: 4, correctAnswer: 5 },
-        { questionNumber: 24, score: 4, correctAnswer: 7 },
-        { questionNumber: 25, score: 4, correctAnswer: 6 },
-        { questionNumber: 26, score: 4, correctAnswer: 1 },
-      ],
-      sectionTotal: { score: 44 },
-    },
-    {
-      section: 3,
-      questions: [
-        { questionNumber: 27, score: 4, correctAnswer: 4 },
-        { questionNumber: 28, score: 4, correctAnswer: 2 },
-        { questionNumber: 29, score: 5, correctAnswer: 2 },
-        { questionNumber: 30, score: 5, correctAnswer: 4 },
-        { questionNumber: 31, score: 5, correctAnswer: 1 },
-        { questionNumber: 32, score: 6, correctAnswer: 2 },
-        { questionNumber: 33, score: 6, correctAnswer: 2 },
-        { questionNumber: 34, score: 6, correctAnswer: 3 },
-      ],
-      sectionTotal: { score: 41 },
-    },
-    {
-      section: 4,
-      questions: [
-        { questionNumber: 35, score: 5, correctAnswer: 4 },
-        { questionNumber: 36, score: 5, correctAnswer: 1 },
-        { questionNumber: 37, score: 5, correctAnswer: 2 },
-        { questionNumber: 38, score: 5, correctAnswer: 2 },
-        { questionNumber: 39, score: 5, correctAnswer: 2 },
-        { questionNumber: 40, score: 5, correctAnswer: 3 },
-        { questionNumber: 41, score: 5, correctAnswer: 1 },
-      ],
-      sectionTotal: { score: 35 },
-    },
-    {
-      section: 5,
-      questions: [
-        { questionNumber: 42, score: 6, correctAnswer: 3 },
-        { questionNumber: 43, score: 6, correctAnswer: 3 },
-        { questionNumber: 44, score: 6, correctAnswer: 4 },
-        { questionNumber: 45, score: 6, correctAnswer: 4 },
-        { questionNumber: 46, score: 6, correctAnswer: 4 },
-      ],
-      sectionTotal: { score: 30 },
-    },
-    {
-      section: 6,
-      questions: [
-        { questionNumber: 47, score: 6, correctAnswer: 4 },
-        { questionNumber: 48, score: 6, correctAnswer: 2 },
-        { questionNumber: 49, score: 6, correctAnswer: 2 },
-        { questionNumber: 50, score: 6, correctAnswer: 2 },
-        { questionNumber: 51, score: 6, correctAnswer: 1 },
-        { questionNumber: 52, score: 6, correctAnswer: 3 },
-        { questionNumber: 53, score: 6, correctAnswer: 1 },
-        { questionNumber: 54, score: 6, correctAnswer: 2 },
-        { questionNumber: 55, score: 6, correctAnswer: 4 },
-      ],
-      sectionTotal: { score: 36 },
-    },
-  ],
-};
-
-const mockTestData_R_2016: TestData = {
-  subject: "Reading",
-  year: 2016,
-  maxScore: 200,
-  testStructure: [
-    {
-      section: 1,
-      questions: [
-        { questionNumber: 1, score: 2, correctAnswer: 2 },
-        { questionNumber: 2, score: 2, correctAnswer: 4 },
-        { questionNumber: 3, score: 2, correctAnswer: 2 },
-        { questionNumber: 4, score: 2, correctAnswer: 3 },
-        { questionNumber: 5, score: 2, correctAnswer: 3 },
-        { questionNumber: 6, score: 2, correctAnswer: 4 },
-        { questionNumber: 7, score: 2, correctAnswer: 1 },
-      ],
-      sectionTotal: { score: 14 },
-    },
-    {
-      section: 2,
-      questions: [
-        { questionNumber: 8, score: 2, correctAnswer: 1 },
-        { questionNumber: 9, score: 2, correctAnswer: 1 },
-        { questionNumber: 10, score: 2, correctAnswer: 2 },
-        { questionNumber: 11, score: 2, correctAnswer: 4 },
-        { questionNumber: 12, score: 2, correctAnswer: 2 },
-        { questionNumber: 13, score: 2, correctAnswer: 1 },
-        { questionNumber: 14, score: 2, correctAnswer: 4 },
-        { questionNumber: 15, score: 2, correctAnswer: 3 },
-        { questionNumber: 16, score: 2, correctAnswer: 1 },
-        { questionNumber: 17, score: 2, correctAnswer: 4 },
-        { questionNumber: 18, score: 4, correctAnswer: 5 },
-        { questionNumber: 19, score: 4, correctAnswer: 1 },
-        { questionNumber: 20, score: 4, correctAnswer: 3 },
-        { questionNumber: 21, score: 4, correctAnswer: 5 },
-        { questionNumber: 22, score: 4, correctAnswer: 4 },
-        { questionNumber: 23, score: 4, correctAnswer: 6 },
-        { questionNumber: 24, score: 4, correctAnswer: 2 },
-        { questionNumber: 25, score: 4, correctAnswer: 2 },
-        { questionNumber: 26, score: 4, correctAnswer: 4 },
-      ],
-      sectionTotal: { score: 44 },
-    },
-    {
-      section: 3,
-      questions: [
-        { questionNumber: 27, score: 4, correctAnswer: 4 },
-        { questionNumber: 28, score: 4, correctAnswer: 2 },
-        { questionNumber: 29, score: 5, correctAnswer: 1 },
-        { questionNumber: 30, score: 5, correctAnswer: 2 },
-        { questionNumber: 31, score: 5, correctAnswer: 1 },
-        { questionNumber: 32, score: 6, correctAnswer: 4 },
-        { questionNumber: 33, score: 6, correctAnswer: 4 },
-        { questionNumber: 34, score: 6, correctAnswer: 1 },
-      ],
-      sectionTotal: { score: 41 },
-    },
-    {
-      section: 4,
-      questions: [
-        { questionNumber: 35, score: 5, correctAnswer: 2 },
-        { questionNumber: 36, score: 5, correctAnswer: 1 },
-        { questionNumber: 37, score: 5, correctAnswer: 3 },
-        { questionNumber: 38, score: 5, correctAnswer: 2 },
-        { questionNumber: 39, score: 5, correctAnswer: 2 },
-        { questionNumber: 40, score: 5, correctAnswer: 3 },
-        { questionNumber: 41, score: 5, correctAnswer: 1 },
-      ],
-      sectionTotal: { score: 35 },
-    },
-    {
-      section: 5,
-      questions: [
-        { questionNumber: 42, score: 6, correctAnswer: 2 },
-        { questionNumber: 43, score: 6, correctAnswer: 3 },
-        { questionNumber: 44, score: 6, correctAnswer: 2 },
-        { questionNumber: 45, score: 6, correctAnswer: 3 },
-        { questionNumber: 46, score: 6, correctAnswer: 2 },
-      ],
-      sectionTotal: { score: 30 },
-    },
-    {
-      section: 6,
-      questions: [
-        { questionNumber: 47, score: 6, correctAnswer: 1 },
-        { questionNumber: 48, score: 6, correctAnswer: 3 },
-        { questionNumber: 49, score: 6, correctAnswer: 1 },
-        { questionNumber: 50, score: 6, correctAnswer: 3 },
-        { questionNumber: 51, score: 6, correctAnswer: 3 },
-        { questionNumber: 52, score: 6, correctAnswer: 3 },
-        { questionNumber: 53, score: 6, correctAnswer: 1 },
-        { questionNumber: 54, score: 6, correctAnswer: 4 },
-        { questionNumber: 55, score: 6, correctAnswer: 2 },
-      ],
-      sectionTotal: { score: 36 },
-    },
-  ],
-};
-
-const mockTestData_R_2017: TestData = {
-  subject: "Reading",
-  year: 2017,
-  maxScore: 200,
-  testStructure: [
-    {
-      section: 1,
-      questions: [
-        { questionNumber: 1, score: 2, correctAnswer: 4 },
-        { questionNumber: 2, score: 2, correctAnswer: 3 },
-        { questionNumber: 3, score: 2, correctAnswer: 4 },
-        { questionNumber: 4, score: 2, correctAnswer: 2 },
-        { questionNumber: 5, score: 2, correctAnswer: 1 },
-        { questionNumber: 6, score: 2, correctAnswer: 2 },
-        { questionNumber: 7, score: 2, correctAnswer: 1 },
-      ],
-      sectionTotal: { score: 14 },
-    },
-    {
-      section: 2,
-      questions: [
-        { questionNumber: 8, score: 2, correctAnswer: 1 },
-        { questionNumber: 9, score: 2, correctAnswer: 2 },
-        { questionNumber: 10, score: 2, correctAnswer: 4 },
-        { questionNumber: 11, score: 2, correctAnswer: 1 },
-        { questionNumber: 12, score: 2, correctAnswer: 2 },
-        { questionNumber: 13, score: 2, correctAnswer: 3 },
-        { questionNumber: 14, score: 2, correctAnswer: 4 },
-        { questionNumber: 15, score: 2, correctAnswer: 1 },
-        { questionNumber: 16, score: 2, correctAnswer: 3 },
-        { questionNumber: 17, score: 2, correctAnswer: 1 },
-        { questionNumber: 18, score: 4, correctAnswer: 2 },
-        { questionNumber: 19, score: 4, correctAnswer: 6 },
-        { questionNumber: 20, score: 4, correctAnswer: 5 },
-        { questionNumber: 21, score: 4, correctAnswer: 1 },
-        { questionNumber: 22, score: 4, correctAnswer: 6 },
-        { questionNumber: 23, score: 4, correctAnswer: 2 },
-        { questionNumber: 24, score: 4, correctAnswer: 5 },
-        { questionNumber: 25, score: 4, correctAnswer: 2 },
-        { questionNumber: 26, score: 4, correctAnswer: 7 },
-      ],
-      sectionTotal: { score: 44 },
-    },
-    {
-      section: 3,
-      questions: [
-        { questionNumber: 27, score: 4, correctAnswer: 2 },
-        { questionNumber: 28, score: 4, correctAnswer: 3 },
-        { questionNumber: 29, score: 5, correctAnswer: 3 },
-        { questionNumber: 30, score: 5, correctAnswer: 2 },
-        { questionNumber: 31, score: 5, correctAnswer: 3 },
-        { questionNumber: 32, score: 6, correctAnswer: 3 },
-        { questionNumber: 33, score: 6, correctAnswer: 4 },
-        { questionNumber: 34, score: 6, correctAnswer: 2 },
-      ],
-      sectionTotal: { score: 41 },
-    },
-    {
-      section: 4,
-      questions: [
-        { questionNumber: 35, score: 5, correctAnswer: 3 },
-        { questionNumber: 36, score: 5, correctAnswer: 1 },
-        { questionNumber: 37, score: 5, correctAnswer: 4 },
-        { questionNumber: 38, score: 5, correctAnswer: 2 },
-        { questionNumber: 39, score: 5, correctAnswer: 4 },
-        { questionNumber: 40, score: 5, correctAnswer: 2 },
-        { questionNumber: 41, score: 5, correctAnswer: 4 },
-      ],
-      sectionTotal: { score: 35 },
-    },
-    {
-      section: 5,
-      questions: [
-        { questionNumber: 42, score: 6, correctAnswer: 1 },
-        { questionNumber: 43, score: 6, correctAnswer: 2 },
-        { questionNumber: 44, score: 6, correctAnswer: 4 },
-        { questionNumber: 45, score: 6, correctAnswer: 1 },
-        { questionNumber: 46, score: 6, correctAnswer: 2 },
-      ],
-      sectionTotal: { score: 30 },
-    },
-    {
-      section: 6,
-      questions: [
-        { questionNumber: 47, score: 6, correctAnswer: 4 },
-        { questionNumber: 48, score: 6, correctAnswer: 2 },
-        { questionNumber: 49, score: 6, correctAnswer: 4 },
-        { questionNumber: 50, score: 6, correctAnswer: 4 },
-        { questionNumber: 51, score: 6, correctAnswer: 1 },
-        { questionNumber: 52, score: 6, correctAnswer: 4 },
-        { questionNumber: 53, score: 6, correctAnswer: 2 },
-        { questionNumber: 54, score: 6, correctAnswer: 3 },
-        { questionNumber: 55, score: 6, correctAnswer: 1 },
-      ],
-      sectionTotal: { score: 36 },
-    },
-  ],
-};
-
-const mockTestData_R_2018: TestData = {
-  subject: "Reading",
-  year: 2018,
-  maxScore: 200,
-  testStructure: [
-    {
-      section: 1,
-      questions: [
-        { questionNumber: 1, score: 2, correctAnswer: 4 },
-        { questionNumber: 2, score: 2, correctAnswer: 3 },
-        { questionNumber: 3, score: 2, correctAnswer: 2 },
-        { questionNumber: 4, score: 2, correctAnswer: 1 },
-        { questionNumber: 5, score: 2, correctAnswer: 4 },
-        { questionNumber: 6, score: 2, correctAnswer: 3 },
-        { questionNumber: 7, score: 2, correctAnswer: 2 },
-      ],
-      sectionTotal: { score: 14 },
-    },
-    {
-      section: 2,
-      questions: [
-        { questionNumber: 8, score: 2, correctAnswer: 3 },
-        { questionNumber: 9, score: 2, correctAnswer: 2 },
-        { questionNumber: 10, score: 2, correctAnswer: 1 },
-        { questionNumber: 11, score: 2, correctAnswer: 4 },
-        { questionNumber: 12, score: 2, correctAnswer: 1 },
-        { questionNumber: 13, score: 2, correctAnswer: 3 },
-        { questionNumber: 14, score: 2, correctAnswer: 1 },
-        { questionNumber: 15, score: 2, correctAnswer: 3 },
-        { questionNumber: 16, score: 2, correctAnswer: 3 },
-        { questionNumber: 17, score: 2, correctAnswer: 3 },
-        { questionNumber: 18, score: 2, correctAnswer: 3 },
-        { questionNumber: 19, score: 4, correctAnswer: 2 },
-        { questionNumber: 20, score: 4, correctAnswer: 4 },
-        { questionNumber: 21, score: 4, correctAnswer: 2 },
-        { questionNumber: 22, score: 4, correctAnswer: 3 },
-        { questionNumber: 23, score: 4, correctAnswer: 2 },
-        { questionNumber: 24, score: 5, correctAnswer: 2 },
-        { questionNumber: 25, score: 5, correctAnswer: 4 },
-        { questionNumber: 26, score: 5, correctAnswer: 8 },
-      ],
-      sectionTotal: { score: 47 },
-    },
-    {
-      section: 3,
-      questions: [
-        { questionNumber: 27, score: 5, correctAnswer: 2 },
-        { questionNumber: 30, score: 6, correctAnswer: 1 },
-        { questionNumber: 31, score: 6, correctAnswer: 3 },
-        { questionNumber: 32, score: 6, correctAnswer: 4 },
-      ],
-      sectionTotal: { score: 23 },
-    },
-    {
-      section: 4,
-      questions: [
-        { questionNumber: 33, score: 5, correctAnswer: 2 },
-        { questionNumber: 34, score: 5, correctAnswer: 3 },
-        { questionNumber: 35, score: 5, correctAnswer: 2 },
-        { questionNumber: 36, score: 5, correctAnswer: 4 },
-        { questionNumber: 37, score: 5, correctAnswer: 4 },
-        { questionNumber: 38, score: 5, correctAnswer: 2 },
-        { questionNumber: 39, score: 5, correctAnswer: 3 },
-        { questionNumber: 40, score: 5, correctAnswer: 4 },
-      ],
-      sectionTotal: { score: 40 },
-    },
-    {
-      section: 5,
-      questions: [
-        { questionNumber: 41, score: 6, correctAnswer: 3 },
-        { questionNumber: 42, score: 6, correctAnswer: 2 },
-        { questionNumber: 43, score: 6, correctAnswer: 2 },
-        { questionNumber: 44, score: 6, correctAnswer: 1 },
-        { questionNumber: 45, score: 6, correctAnswer: 3 },
-      ],
-      sectionTotal: { score: 30 },
-    },
-    {
-      section: 6,
-      questions: [
-        { questionNumber: 46, score: 6, correctAnswer: 4 },
-        { questionNumber: 47, score: 6, correctAnswer: 2 },
-        { questionNumber: 48, score: 6, correctAnswer: 1 },
-        { questionNumber: 49, score: 6, correctAnswer: 1 },
-        { questionNumber: 50, score: 6, correctAnswer: 2 },
-        { questionNumber: 51, score: 6, correctAnswer: 4 },
-        { questionNumber: 52, score: null, correctAnswer: 2 },
-        { questionNumber: 53, score: null, correctAnswer: 3 },
-        { questionNumber: 54, score: null, correctAnswer: 1 },
-      ],
-      sectionTotal: { score: 36 },
-    },
-  ],
-};
-
-const mockTestData_R_2019: TestData = {
-  subject: "Reading",
-  year: 2019,
-  maxScore: 200,
-  testStructure: [
-    {
-      section: 1,
-      questions: [
-        { questionNumber: 1, score: 2, correctAnswer: 2 },
-        { questionNumber: 2, score: 2, correctAnswer: 1 },
-        { questionNumber: 3, score: 2, correctAnswer: 2 },
-        { questionNumber: 4, score: 2, correctAnswer: 3 },
-        { questionNumber: 5, score: 2, correctAnswer: 2 },
-        { questionNumber: 6, score: 2, correctAnswer: 2 },
-        { questionNumber: 7, score: 2, correctAnswer: 1 },
-      ],
-      sectionTotal: { score: 14 },
-    },
-    {
-      section: 2,
-      questions: [
-        { questionNumber: 8, score: 2, correctAnswer: 2 },
-        { questionNumber: 9, score: 2, correctAnswer: 3 },
-        { questionNumber: 10, score: 2, correctAnswer: 1 },
-        { questionNumber: 11, score: 2, correctAnswer: 2 },
-        { questionNumber: 12, score: 2, correctAnswer: 4 },
-        { questionNumber: 13, score: 2, correctAnswer: 3 },
-        { questionNumber: 14, score: 2, correctAnswer: 3 },
-        { questionNumber: 15, score: 2, correctAnswer: 2 },
-        { questionNumber: 16, score: 2, correctAnswer: 4 },
-        { questionNumber: 17, score: 2, correctAnswer: 4 },
-        { questionNumber: 18, score: 4, correctAnswer: 2 },
-        { questionNumber: 19, score: 4, correctAnswer: 5 },
-        { questionNumber: 20, score: 4, correctAnswer: 6 },
-        { questionNumber: 21, score: 4, correctAnswer: 2 },
-        { questionNumber: 22, score: 4, correctAnswer: 2 },
-        { questionNumber: 23, score: 4, correctAnswer: 6 },
-        { questionNumber: 24, score: 5, correctAnswer: 6 },
-        { questionNumber: 25, score: 5, correctAnswer: 2 },
-        { questionNumber: 26, score: 5, correctAnswer: 3 },
-      ],
-      sectionTotal: { score: 47 },
-    },
-    {
-      section: 3,
-      questions: [
-        { questionNumber: 27, score: 5, correctAnswer: 1 },
-        { questionNumber: 28, score: 5, correctAnswer: 2 },
-        { questionNumber: 29, score: 5, correctAnswer: 4 },
-        { questionNumber: 30, score: 6, correctAnswer: 1 },
-        { questionNumber: 31, score: 6, correctAnswer: 3 },
-        { questionNumber: 32, score: 6, correctAnswer: 3 },
-      ],
-      sectionTotal: { score: 33 },
-    },
-    {
-      section: 4,
-      questions: [
-        { questionNumber: 33, score: 5, correctAnswer: 2 },
-        { questionNumber: 34, score: 5, correctAnswer: 4 },
-        { questionNumber: 35, score: 5, correctAnswer: 2 },
-        { questionNumber: 36, score: 5, correctAnswer: 3 },
-        { questionNumber: 37, score: 5, correctAnswer: 3 },
-        { questionNumber: 38, score: 5, correctAnswer: 2 },
-        { questionNumber: 39, score: 5, correctAnswer: 2 },
-        { questionNumber: 40, score: 5, correctAnswer: 2 },
-      ],
-      sectionTotal: { score: 40 },
-    },
-    {
-      section: 5,
-      questions: [
-        { questionNumber: 41, score: 6, correctAnswer: 1 },
-        { questionNumber: 42, score: 6, correctAnswer: 2 },
-        { questionNumber: 43, score: 6, correctAnswer: 1 },
-        { questionNumber: 44, score: 6, correctAnswer: 3 },
-        { questionNumber: 45, score: 6, correctAnswer: 3 },
-      ],
-      sectionTotal: { score: 30 },
-    },
-    {
-      section: 6,
-      questions: [
-        { questionNumber: 46, score: 6, correctAnswer: 2 },
-        { questionNumber: 47, score: 6, correctAnswer: 3 },
-        { questionNumber: 48, score: 6, correctAnswer: 4 },
-        { questionNumber: 49, score: 6, correctAnswer: 1 },
-        { questionNumber: 50, score: 6, correctAnswer: 4 },
-        { questionNumber: 51, score: 6, correctAnswer: 1 },
-        { questionNumber: 52, score: 6, correctAnswer: 4 },
-        { questionNumber: 53, score: 6, correctAnswer: 2 },
-        { questionNumber: 54, score: 6, correctAnswer: 3 },
-      ],
-      sectionTotal: { score: 36 },
-    },
-  ],
-};
-
-const mockTestData_R_2020: TestData = {
-  subject: "Reading",
-  year: 2020,
-  maxScore: 200,
-  testStructure: [
-    {
-      section: 1,
-      questions: [
-        { questionNumber: 1, score: 2, correctAnswer: 2 },
-        { questionNumber: 2, score: 2, correctAnswer: 3 },
-        { questionNumber: 3, score: 2, correctAnswer: 4 },
-        { questionNumber: 4, score: 2, correctAnswer: 2 },
-        { questionNumber: 5, score: 2, correctAnswer: 1 },
-        { questionNumber: 6, score: 2, correctAnswer: 1 },
-        { questionNumber: 7, score: 2, correctAnswer: 2 },
-      ],
-      sectionTotal: { score: 14 },
-    },
-    {
-      section: 2,
-      questions: [
-        { questionNumber: 8, score: 2, correctAnswer: 3 },
-        { questionNumber: 9, score: 2, correctAnswer: 1 },
-        { questionNumber: 10, score: 2, correctAnswer: 3 },
-        { questionNumber: 11, score: 2, correctAnswer: 3 },
-        { questionNumber: 12, score: 2, correctAnswer: 1 },
-        { questionNumber: 13, score: 2, correctAnswer: 1 },
-        { questionNumber: 14, score: 2, correctAnswer: 1 },
-        { questionNumber: 15, score: 2, correctAnswer: 4 },
-        { questionNumber: 16, score: 2, correctAnswer: 2 },
-        { questionNumber: 17, score: 2, correctAnswer: 1 },
-        { questionNumber: 18, score: 4, correctAnswer: 4 },
-        { questionNumber: 19, score: 4, correctAnswer: 2 },
-        { questionNumber: 20, score: 4, correctAnswer: 4 },
-        { questionNumber: 21, score: 4, correctAnswer: 3 },
-        { questionNumber: 22, score: 4, correctAnswer: 5 },
-        { questionNumber: 23, score: 4, correctAnswer: 2 },
-        { questionNumber: 24, score: 5, correctAnswer: 3 },
-        { questionNumber: 25, score: 5, correctAnswer: 1 },
-        { questionNumber: 26, score: 5, correctAnswer: 3 },
-      ],
-      sectionTotal: { score: 47 },
-    },
-    {
-      section: 3,
-      questions: [
-        { questionNumber: 27, score: 5, correctAnswer: 2 },
-        { questionNumber: 28, score: 5, correctAnswer: 2 },
-        { questionNumber: 29, score: 5, correctAnswer: 2 },
-        { questionNumber: 30, score: 6, correctAnswer: 3 },
-        { questionNumber: 31, score: 6, correctAnswer: 3 },
-        { questionNumber: 32, score: 6, correctAnswer: 4 },
-      ],
-      sectionTotal: { score: 33 },
-    },
-    {
-      section: 4,
-      questions: [
-        { questionNumber: 33, score: 5, correctAnswer: 4 },
-        { questionNumber: 34, score: 5, correctAnswer: 4 },
-        { questionNumber: 35, score: 5, correctAnswer: 2 },
-        { questionNumber: 36, score: 5, correctAnswer: 4 },
-        { questionNumber: 37, score: 5, correctAnswer: 1 },
-        { questionNumber: 38, score: 5, correctAnswer: 1 },
-        { questionNumber: 39, score: 5, correctAnswer: 2 },
-        { questionNumber: 40, score: 5, correctAnswer: 4 },
-      ],
-      sectionTotal: { score: 40 },
-    },
-    {
-      section: 5,
-      questions: [
-        { questionNumber: 41, score: 6, correctAnswer: 1 },
-        { questionNumber: 42, score: 6, correctAnswer: 4 },
-        { questionNumber: 43, score: 6, correctAnswer: 4 },
-        { questionNumber: 44, score: 6, correctAnswer: 2 },
-        { questionNumber: 45, score: 6, correctAnswer: 2 },
-      ],
-      sectionTotal: { score: 30 },
-    },
-    {
-      section: 6,
-      questions: [
-        { questionNumber: 46, score: 6, correctAnswer: 1 },
-        { questionNumber: 47, score: 6, correctAnswer: 3 },
-        { questionNumber: 48, score: 6, correctAnswer: 2 },
-        { questionNumber: 49, score: 6, correctAnswer: 4 },
-        { questionNumber: 50, score: 6, correctAnswer: 2 },
-        { questionNumber: 51, score: 6, correctAnswer: 2 },
-        { questionNumber: 52, score: 6, correctAnswer: 4 },
-        { questionNumber: 53, score: 6, correctAnswer: 3 },
-        { questionNumber: 54, score: 6, correctAnswer: 1 },
-      ],
-      sectionTotal: { score: 36 },
-    },
-  ],
-};
-
-const mockTestData_1A_2015: TestData = {
-  subject: "Math1A",
-  year: 2015,
-  maxScore: 100,
-  testStructure: [
-    {
-      section: 1,
-      questions: [
-        { questionNumber: 1, score: 5, correctAnswer: null },
-        { questionNumber: 2, score: 5, correctAnswer: null },
-        { questionNumber: 3, score: 5, correctAnswer: null },
-        { questionNumber: 4, score: 2, correctAnswer: null },
-        { questionNumber: 5, score: 3, correctAnswer: null },
-      ],
-      sectionTotal: { score: 20 },
-    },
-    {
-      section: 2,
-      questions: [
-        { questionNumber: 6, score: 4, correctAnswer: null },
-        { questionNumber: 7, score: 3, correctAnswer: null },
-        { questionNumber: 8, score: 3, correctAnswer: null },
-        { questionNumber: 9, score: 3, correctAnswer: null },
-        { questionNumber: 10, score: 3, correctAnswer: null },
-        { questionNumber: 11, score: 3, correctAnswer: null },
-        { questionNumber: 12, score: 3, correctAnswer: null },
-        { questionNumber: 13, score: 3, correctAnswer: null },
-      ],
-      sectionTotal: { score: 25 },
-    },
-    {
-      section: 3,
-      questions: [
-        { questionNumber: 14, score: 3, correctAnswer: null },
-        { questionNumber: 15, score: 4, correctAnswer: null },
-        { questionNumber: 16, score: 6, correctAnswer: null },
-        { questionNumber: 17, score: 2, correctAnswer: null },
-      ],
-      sectionTotal: { score: 15 },
-    },
-    {
-      section: 4,
-      questions: [
-        { questionNumber: 18, score: 3, correctAnswer: null },
-        { questionNumber: 19, score: 2, correctAnswer: null },
-        { questionNumber: 20, score: 3, correctAnswer: null },
-        { questionNumber: 21, score: 3, correctAnswer: null },
-        { questionNumber: 22, score: 2, correctAnswer: null },
-        { questionNumber: 23, score: 2, correctAnswer: null },
-        { questionNumber: 24, score: 2, correctAnswer: null },
-        { questionNumber: 25, score: 3, correctAnswer: null },
-      ],
-      sectionTotal: { score: 20 },
-    },
-    {
-      section: 5,
-      questions: [
-        { questionNumber: 26, score: 3, correctAnswer: null },
-        { questionNumber: 27, score: 3, correctAnswer: null },
-        { questionNumber: 28, score: 3, correctAnswer: null },
-        { questionNumber: 29, score: 3, correctAnswer: null },
-        { questionNumber: 30, score: 2, correctAnswer: null },
-        { questionNumber: 31, score: 2, correctAnswer: null },
-        { questionNumber: 32, score: 4, correctAnswer: null },
-      ],
-      sectionTotal: { score: 20 },
-    },
-    {
-      section: 6,
-      questions: [
-        { questionNumber: 33, score: 3, correctAnswer: null },
-        { questionNumber: 34, score: 3, correctAnswer: null },
-        { questionNumber: 35, score: 3, correctAnswer: null },
-        { questionNumber: 36, score: 4, correctAnswer: null },
-        { questionNumber: 37, score: 4, correctAnswer: null },
-        { questionNumber: 38, score: 3, correctAnswer: null },
-      ],
-      sectionTotal: { score: 20 },
-    },
-  ],
-};
-
-// 2016年 数学I・数学A テストデータ
-const mockTestData_1A_2016: TestData = {
-  subject: "Math1A",
-  year: 2016,
-  maxScore: 100,
-  testStructure: [
-    {
-      section: 1,
-      questions: [
-        { questionNumber: 1, score: 2, correctAnswer: null },
-        { questionNumber: 2, score: 2, correctAnswer: null },
-        { questionNumber: 3, score: 2, correctAnswer: null },
-        { questionNumber: 4, score: 2, correctAnswer: null },
-        { questionNumber: 5, score: 2, correctAnswer: null },
-        { questionNumber: 6, score: 2, correctAnswer: null },
-        { questionNumber: 7, score: 2, correctAnswer: null },
-        { questionNumber: 8, score: 3, correctAnswer: null },
-        { questionNumber: 9, score: 3, correctAnswer: null },
-        { questionNumber: 10, score: 3, correctAnswer: null },
-        { questionNumber: 11, score: 3, correctAnswer: null },
-        { questionNumber: 12, score: 4, correctAnswer: null },
-      ],
-      sectionTotal: { score: 30 },
-    },
-    {
-      section: 2,
-      questions: [
-        { questionNumber: 13, score: 3, correctAnswer: null },
-        { questionNumber: 14, score: 3, correctAnswer: null },
-        { questionNumber: 15, score: 3, correctAnswer: null },
-        { questionNumber: 16, score: 3, correctAnswer: null },
-        { questionNumber: 17, score: 3, correctAnswer: null },
-        { questionNumber: 18, score: 3, correctAnswer: null },
-        { questionNumber: 19, score: 3, correctAnswer: null },
-        { questionNumber: 20, score: 3, correctAnswer: null },
-        { questionNumber: 21, score: 2, correctAnswer: null },
-        { questionNumber: 22, score: 2, correctAnswer: null },
-        { questionNumber: 23, score: 2, correctAnswer: null },
-      ],
-      sectionTotal: { score: 30 },
-    },
-    {
-      section: 3,
-      questions: [
-        { questionNumber: 24, score: 3, correctAnswer: null },
-        { questionNumber: 25, score: 3, correctAnswer: null },
-        { questionNumber: 26, score: 3, correctAnswer: null },
-        { questionNumber: 27, score: 3, correctAnswer: null },
-        { questionNumber: 28, score: 4, correctAnswer: null },
-        { questionNumber: 29, score: 4, correctAnswer: null },
-      ],
-      sectionTotal: { score: 20 },
-    },
-    {
-      section: 4,
-      questions: [
-        { questionNumber: 30, score: 3, correctAnswer: null },
-        { questionNumber: 31, score: 3, correctAnswer: null },
-        { questionNumber: 32, score: 2, correctAnswer: null },
-        { questionNumber: 33, score: 2, correctAnswer: null },
-        { questionNumber: 34, score: 4, correctAnswer: null },
-        { questionNumber: 35, score: 6, correctAnswer: null },
-      ],
-      sectionTotal: { score: 20 },
-    },
-    {
-      section: 5,
-      questions: [
-        { questionNumber: 36, score: 2, correctAnswer: null },
-        { questionNumber: 37, score: 3, correctAnswer: null },
-        { questionNumber: 38, score: 3, correctAnswer: null },
-        { questionNumber: 39, score: 3, correctAnswer: null },
-        { questionNumber: 40, score: 3, correctAnswer: null },
-        { questionNumber: 41, score: 2, correctAnswer: null },
-        { questionNumber: 42, score: 2, correctAnswer: null },
-        { questionNumber: 43, score: 2, correctAnswer: null },
-      ],
-      sectionTotal: { score: 20 },
-    },
-  ],
-};
-
-// 2017年 数学I・数学A テストデータ
-const mockTestData_1A_2017: TestData = {
-  subject: "Math1A",
-  year: 2017,
-  maxScore: 100,
-  testStructure: [
-    {
-      section: 1,
-      questions: [
-        { questionNumber: 1, score: 3, correctAnswer: null },
-        { questionNumber: 2, score: 1, correctAnswer: null },
-        { questionNumber: 3, score: 3, correctAnswer: null },
-        { questionNumber: 4, score: 3, correctAnswer: null },
-        { questionNumber: 5, score: 1, correctAnswer: null },
-        { questionNumber: 6, score: 2, correctAnswer: null },
-        { questionNumber: 7, score: 2, correctAnswer: null },
-        { questionNumber: 8, score: 2, correctAnswer: null },
-        { questionNumber: 9, score: 3, correctAnswer: null },
-        { questionNumber: 10, score: 2, correctAnswer: null },
-        { questionNumber: 11, score: 2, correctAnswer: null },
-        { questionNumber: 12, score: 3, correctAnswer: null },
-        { questionNumber: 13, score: 3, correctAnswer: null },
-      ],
-      sectionTotal: { score: 30 },
-    },
-    {
-      section: 2,
-      questions: [
-        { questionNumber: 14, score: 3, correctAnswer: null },
-        { questionNumber: 15, score: 3, correctAnswer: null },
-        { questionNumber: 16, score: 3, correctAnswer: null },
-        { questionNumber: 17, score: 3, correctAnswer: null },
-        { questionNumber: 18, score: 3, correctAnswer: null },
-        { questionNumber: 19, score: 6, correctAnswer: null },
-        { questionNumber: 20, score: 2, correctAnswer: null },
-        { questionNumber: 21, score: 2, correctAnswer: null },
-        { questionNumber: 22, score: 2, correctAnswer: null },
-        { questionNumber: 23, score: 1, correctAnswer: null },
-        { questionNumber: 24, score: 2, correctAnswer: null },
-      ],
-      sectionTotal: { score: 30 },
-    },
-    {
-      section: 3,
-      questions: [
-        { questionNumber: 25, score: 2, correctAnswer: null },
-        { questionNumber: 26, score: 3, correctAnswer: null },
-        { questionNumber: 27, score: 2, correctAnswer: null },
-        { questionNumber: 28, score: 2, correctAnswer: null },
-        { questionNumber: 29, score: 3, correctAnswer: null },
-        { questionNumber: 30, score: 2, correctAnswer: null },
-        { questionNumber: 31, score: 2, correctAnswer: null },
-        { questionNumber: 32, score: 4, correctAnswer: null },
-      ],
-      sectionTotal: { score: 20 },
-    },
-    {
-      section: 4,
-      questions: [
-        { questionNumber: 33, score: 2, correctAnswer: null },
-        { questionNumber: 34, score: 2, correctAnswer: null },
-        { questionNumber: 35, score: 2, correctAnswer: null },
-        { questionNumber: 36, score: 2, correctAnswer: null },
-        { questionNumber: 37, score: 3, correctAnswer: null },
-        { questionNumber: 38, score: 2, correctAnswer: null },
-        { questionNumber: 39, score: 2, correctAnswer: null },
-        { questionNumber: 40, score: 2, correctAnswer: null },
-        { questionNumber: 41, score: 3, correctAnswer: null },
-      ],
-      sectionTotal: { score: 20 },
-    },
-    {
-      section: 5,
-      questions: [
-        { questionNumber: 42, score: 3, correctAnswer: null },
-        { questionNumber: 43, score: 3, correctAnswer: null },
-        { questionNumber: 44, score: 3, correctAnswer: null },
-        { questionNumber: 45, score: 3, correctAnswer: null },
-        { questionNumber: 46, score: 2, correctAnswer: null },
-        { questionNumber: 47, score: 3, correctAnswer: null },
-        { questionNumber: 48, score: 3, correctAnswer: null },
-      ],
-      sectionTotal: { score: 20 },
-    },
-  ],
-};
-
-// 2018年 数学I・数学A テストデータ
-const mockTestData_1A_2018: TestData = {
-  subject: "Math1A",
-  year: 2018,
-  maxScore: 100,
-  testStructure: [
-    {
-      section: 1,
-      questions: [
-        { questionNumber: 1, score: 2, correctAnswer: null },
-        { questionNumber: 2, score: 4, correctAnswer: null },
-        { questionNumber: 3, score: 2, correctAnswer: null },
-        { questionNumber: 4, score: 2, correctAnswer: null },
-        { questionNumber: 5, score: 3, correctAnswer: null },
-        { questionNumber: 6, score: 3, correctAnswer: null },
-        { questionNumber: 7, score: 2, correctAnswer: null },
-        { questionNumber: 8, score: 2, correctAnswer: null },
-        { questionNumber: 9, score: 2, correctAnswer: null },
-        { questionNumber: 10, score: 2, correctAnswer: null },
-        { questionNumber: 11, score: 2, correctAnswer: null },
-        { questionNumber: 12, score: 2, correctAnswer: null },
-        { questionNumber: 13, score: 2, correctAnswer: null },
-      ],
-      sectionTotal: { score: 30 },
-    },
-    {
-      section: 2,
-      questions: [
-        { questionNumber: 14, score: 3, correctAnswer: null },
-        { questionNumber: 15, score: 3, correctAnswer: null },
-        { questionNumber: 16, score: 5, correctAnswer: null },
-        { questionNumber: 17, score: 4, correctAnswer: null },
-        { questionNumber: 18, score: 6, correctAnswer: null },
-        { questionNumber: 19, score: 6, correctAnswer: null },
-        { questionNumber: 20, score: 3, correctAnswer: null },
-      ],
-      sectionTotal: { score: 30 },
-    },
-    {
-      section: 3,
-      questions: [
-        { questionNumber: 21, score: 2, correctAnswer: null },
-        { questionNumber: 22, score: 2, correctAnswer: null },
-        { questionNumber: 23, score: 2, correctAnswer: null },
-        { questionNumber: 24, score: 2, correctAnswer: null },
-        { questionNumber: 25, score: 2, correctAnswer: null },
-        { questionNumber: 26, score: 2, correctAnswer: null },
-        { questionNumber: 27, score: 2, correctAnswer: null },
-        { questionNumber: 28, score: 3, correctAnswer: null },
-        { questionNumber: 29, score: 3, correctAnswer: null },
-      ],
-      sectionTotal: { score: 20 },
-    },
-    {
-      section: 4,
-      questions: [
-        { questionNumber: 30, score: 3, correctAnswer: null },
-        { questionNumber: 31, score: 3, correctAnswer: null },
-        { questionNumber: 32, score: 2, correctAnswer: null },
-        { questionNumber: 33, score: 2, correctAnswer: null },
-        { questionNumber: 34, score: 2, correctAnswer: null },
-        { questionNumber: 35, score: 2, correctAnswer: null },
-        { questionNumber: 36, score: 3, correctAnswer: null },
-        { questionNumber: 37, score: 3, correctAnswer: null },
-      ],
-      sectionTotal: { score: 20 },
-    },
-    {
-      section: 5,
-      questions: [
-        { questionNumber: 38, score: 3, correctAnswer: null },
-        { questionNumber: 39, score: 3, correctAnswer: null },
-        { questionNumber: 40, score: 2, correctAnswer: null },
-        { questionNumber: 41, score: 4, correctAnswer: null },
-        { questionNumber: 42, score: 3, correctAnswer: null },
-        { questionNumber: 43, score: 2, correctAnswer: null },
-        { questionNumber: 44, score: 3, correctAnswer: null },
-      ],
-      sectionTotal: { score: 20 },
-    },
-  ],
-};
-
-// 2019年 数学I・数学A テストデータ
-const mockTestData_1A_2019: TestData = {
-  subject: "Math1A",
-  year: 2019,
-  maxScore: 100,
-  testStructure: [
-    {
-      section: 1,
-      questions: [
-        { questionNumber: 1, score: 2, correctAnswer: null },
-        { questionNumber: 2, score: 2, correctAnswer: null },
-        { questionNumber: 3, score: 2, correctAnswer: null },
-        { questionNumber: 4, score: 2, correctAnswer: null },
-        { questionNumber: 5, score: 2, correctAnswer: null },
-        { questionNumber: 6, score: 2, correctAnswer: null },
-        { questionNumber: 7, score: 2, correctAnswer: null },
-        { questionNumber: 8, score: 2, correctAnswer: null },
-        { questionNumber: 9, score: 2, correctAnswer: null },
-        { questionNumber: 10, score: 2, correctAnswer: null },
-        { questionNumber: 11, score: 2, correctAnswer: null },
-        { questionNumber: 12, score: 2, correctAnswer: null },
-        { questionNumber: 13, score: 2, correctAnswer: null },
-        { questionNumber: 14, score: 2, correctAnswer: null },
-        { questionNumber: 15, score: 2, correctAnswer: null },
-      ],
-      sectionTotal: { score: 30 },
-    },
-    {
-      section: 2,
-      questions: [
-        { questionNumber: 16, score: 4, correctAnswer: null },
-        { questionNumber: 17, score: 3, correctAnswer: null },
-        { questionNumber: 18, score: 2, correctAnswer: null },
-        { questionNumber: 19, score: 3, correctAnswer: null },
-        { questionNumber: 20, score: 3, correctAnswer: null },
-        { questionNumber: 21, score: 3, correctAnswer: null },
-        { questionNumber: 22, score: 3, correctAnswer: null },
-        { questionNumber: 23, score: 4, correctAnswer: null },
-        { questionNumber: 24, score: 1, correctAnswer: null },
-        { questionNumber: 25, score: 1, correctAnswer: null },
-        { questionNumber: 26, score: 1, correctAnswer: null },
-        { questionNumber: 27, score: 2, correctAnswer: null },
-      ],
-      sectionTotal: { score: 30 },
-    },
-    {
-      section: 3,
-      questions: [
-        { questionNumber: 28, score: 2, correctAnswer: null },
-        { questionNumber: 29, score: 2, correctAnswer: null },
-        { questionNumber: 30, score: 3, correctAnswer: null },
-        { questionNumber: 31, score: 2, correctAnswer: null },
-        { questionNumber: 32, score: 2, correctAnswer: null },
-        { questionNumber: 33, score: 3, correctAnswer: null },
-        { questionNumber: 34, score: 3, correctAnswer: null },
-        { questionNumber: 35, score: 3, correctAnswer: null },
-      ],
-      sectionTotal: { score: 20 },
-    },
-    {
-      section: 4,
-      questions: [
-        { questionNumber: 36, score: 3, correctAnswer: null },
-        { questionNumber: 37, score: 2, correctAnswer: null },
-        { questionNumber: 38, score: 3, correctAnswer: null },
-        { questionNumber: 39, score: 3, correctAnswer: null },
-        { questionNumber: 40, score: 2, correctAnswer: null },
-        { questionNumber: 41, score: 2, correctAnswer: null },
-        { questionNumber: 42, score: 2, correctAnswer: null },
-        { questionNumber: 43, score: 3, correctAnswer: null },
-      ],
-      sectionTotal: { score: 20 },
-    },
-    {
-      section: 5,
-      questions: [
-        { questionNumber: 44, score: 4, correctAnswer: null },
-        { questionNumber: 45, score: 3, correctAnswer: null },
-        { questionNumber: 46, score: 3, correctAnswer: null },
-        { questionNumber: 47, score: 2, correctAnswer: null },
-        { questionNumber: 48, score: 2, correctAnswer: null },
-        { questionNumber: 49, score: 3, correctAnswer: null },
-        { questionNumber: 50, score: 3, correctAnswer: null },
-      ],
-      sectionTotal: { score: 20 },
-    },
-  ],
-};
-
-// 2020年 数学I・数学A テストデータ
-const mockTestData_1A_2020: TestData = {
-  subject: "Math1A",
-  year: 2020,
-  maxScore: 100,
-  testStructure: [
-    {
-      section: 1,
-      questions: [
-        { questionNumber: 1, score: 3, correctAnswer: null },
-        { questionNumber: 2, score: 2, correctAnswer: null },
-        { questionNumber: 3, score: 2, correctAnswer: null },
-        { questionNumber: 4, score: 3, correctAnswer: null },
-        { questionNumber: 5, score: 2, correctAnswer: null },
-        { questionNumber: 6, score: 2, correctAnswer: null },
-        { questionNumber: 7, score: 2, correctAnswer: null },
-        { questionNumber: 8, score: 2, correctAnswer: null },
-        { questionNumber: 9, score: 2, correctAnswer: null },
-        { questionNumber: 10, score: 2, correctAnswer: null },
-        { questionNumber: 11, score: 2, correctAnswer: null },
-        { questionNumber: 12, score: 2, correctAnswer: null },
-        { questionNumber: 13, score: 2, correctAnswer: null },
-        { questionNumber: 14, score: 2, correctAnswer: null },
-      ],
-      sectionTotal: { score: 30 },
-    },
-    {
-      section: 2,
-      questions: [
-        { questionNumber: 15, score: 3, correctAnswer: null },
-        { questionNumber: 16, score: 3, correctAnswer: null },
-        { questionNumber: 17, score: 3, correctAnswer: null },
-        { questionNumber: 18, score: 3, correctAnswer: null },
-        { questionNumber: 19, score: 3, correctAnswer: null },
-        { questionNumber: 20, score: 6, correctAnswer: null },
-        { questionNumber: 21, score: 3, correctAnswer: null },
-        { questionNumber: 22, score: 3, correctAnswer: null },
-        { questionNumber: 23, score: 3, correctAnswer: null },
-      ],
-      sectionTotal: { score: 30 },
-    },
-    {
-      section: 3,
-      questions: [
-        { questionNumber: 24, score: 4, correctAnswer: null },
-        { questionNumber: 25, score: 2, correctAnswer: null },
-        { questionNumber: 26, score: 2, correctAnswer: null },
-        { questionNumber: 27, score: 2, correctAnswer: null },
-        { questionNumber: 28, score: 3, correctAnswer: null },
-        { questionNumber: 29, score: 4, correctAnswer: null },
-        { questionNumber: 30, score: 3, correctAnswer: null },
-      ],
-      sectionTotal: { score: 20 },
-    },
-    {
-      section: 4,
-      questions: [
-        { questionNumber: 31, score: 3, correctAnswer: null },
-        { questionNumber: 32, score: 3, correctAnswer: null },
-        { questionNumber: 33, score: 2, correctAnswer: null },
-        { questionNumber: 34, score: 2, correctAnswer: null },
-        { questionNumber: 35, score: 3, correctAnswer: null },
-        { questionNumber: 36, score: 3, correctAnswer: null },
-        { questionNumber: 37, score: 4, correctAnswer: null },
-      ],
-      sectionTotal: { score: 20 },
-    },
-    {
-      section: 5,
-      questions: [
-        { questionNumber: 38, score: 2, correctAnswer: null },
-        { questionNumber: 39, score: 2, correctAnswer: null },
-        { questionNumber: 40, score: 2, correctAnswer: null },
-        { questionNumber: 41, score: 4, correctAnswer: null },
-        { questionNumber: 42, score: 4, correctAnswer: null },
-        { questionNumber: 43, score: 2, correctAnswer: null },
-        { questionNumber: 44, score: 4, correctAnswer: null },
-      ],
-      sectionTotal: { score: 20 },
-    },
-  ],
-};
-
-// 2015年 数学II・数学B（旧数学II・旧数学B）テストデータ
-const mockTestData_2B_2015: TestData = {
-  subject: "Math2B",
-  year: 2015,
-  maxScore: 100,
-  testStructure: [
-    {
-      section: 1,
-      questions: [
-        { questionNumber: 1, score: 1, correctAnswer: null },
-        { questionNumber: 2, score: 1, correctAnswer: null },
-        { questionNumber: 3, score: 1, correctAnswer: null },
-        { questionNumber: 4, score: 1, correctAnswer: null },
-        { questionNumber: 5, score: 2, correctAnswer: null },
-        { questionNumber: 6, score: 2, correctAnswer: null },
-        { questionNumber: 7, score: 1, correctAnswer: null },
-        { questionNumber: 8, score: 2, correctAnswer: null },
-        { questionNumber: 9, score: 1, correctAnswer: null },
-        { questionNumber: 10, score: 3, correctAnswer: null },
-        { questionNumber: 11, score: 3, correctAnswer: null },
-        { questionNumber: 12, score: 3, correctAnswer: null },
-        { questionNumber: 13, score: 2, correctAnswer: null },
-        { questionNumber: 14, score: 2, correctAnswer: null },
-        { questionNumber: 15, score: 2, correctAnswer: null },
-        { questionNumber: 16, score: 3, correctAnswer: null },
-      ],
-      sectionTotal: { score: 30 },
-    },
-    {
-      section: 2,
-      questions: [
-        { questionNumber: 17, score: 2, correctAnswer: null },
-        { questionNumber: 18, score: 2, correctAnswer: null },
-        { questionNumber: 19, score: 1, correctAnswer: null },
-        { questionNumber: 20, score: 3, correctAnswer: null },
-        { questionNumber: 21, score: 1, correctAnswer: null },
-        { questionNumber: 22, score: 3, correctAnswer: null },
-        { questionNumber: 23, score: 4, correctAnswer: null },
-        { questionNumber: 24, score: 5, correctAnswer: null },
-        { questionNumber: 25, score: 2, correctAnswer: null },
-        { questionNumber: 26, score: 2, correctAnswer: null },
-        { questionNumber: 27, score: 2, correctAnswer: null },
-        { questionNumber: 28, score: 3, correctAnswer: null },
-      ],
-      sectionTotal: { score: 30 },
-    },
-    {
-      section: 3,
-      questions: [
-        { questionNumber: 29, score: 2, correctAnswer: null },
-        { questionNumber: 30, score: 1, correctAnswer: null },
-        { questionNumber: 31, score: 2, correctAnswer: null },
-        { questionNumber: 32, score: 2, correctAnswer: null },
-        { questionNumber: 33, score: 1, correctAnswer: null },
-        { questionNumber: 34, score: 1, correctAnswer: null },
-        { questionNumber: 35, score: 1, correctAnswer: null },
-        { questionNumber: 36, score: 3, correctAnswer: null },
-        { questionNumber: 37, score: 2, correctAnswer: null },
-        { questionNumber: 38, score: 3, correctAnswer: null },
-        { questionNumber: 39, score: 1, correctAnswer: null },
-      ],
-      sectionTotal: { score: 20 },
-    },
-    {
-      section: 4,
-      questions: [
-        { questionNumber: 40, score: 2, correctAnswer: null },
-        { questionNumber: 41, score: 1, correctAnswer: null },
-        { questionNumber: 42, score: 1, correctAnswer: null },
-        { questionNumber: 43, score: 1, correctAnswer: null },
-        { questionNumber: 44, score: 1, correctAnswer: null },
-        { questionNumber: 45, score: 2, correctAnswer: null },
-        { questionNumber: 46, score: 1, correctAnswer: null },
-        { questionNumber: 47, score: 1, correctAnswer: null },
-        { questionNumber: 48, score: 2, correctAnswer: null },
-        { questionNumber: 49, score: 2, correctAnswer: null },
-        { questionNumber: 50, score: 2, correctAnswer: null },
-        { questionNumber: 51, score: 2, correctAnswer: null },
-        { questionNumber: 52, score: 2, correctAnswer: null },
-      ],
-      sectionTotal: { score: 20 },
-    },
-    {
-      section: 5,
-      questions: [
-        { questionNumber: 53, score: 3, correctAnswer: null },
-        { questionNumber: 54, score: 2, correctAnswer: null },
-        { questionNumber: 55, score: 2, correctAnswer: null },
-        { questionNumber: 56, score: 2, correctAnswer: null },
-        { questionNumber: 57, score: 2, correctAnswer: null },
-        { questionNumber: 58, score: 2, correctAnswer: null },
-        { questionNumber: 59, score: 2, correctAnswer: null },
-        { questionNumber: 60, score: 2, correctAnswer: null },
-        { questionNumber: 61, score: 1, correctAnswer: null },
-        { questionNumber: 62, score: 2, correctAnswer: null },
-      ],
-      sectionTotal: { score: 20 },
-    },
-    {
-      section: 6,
-      questions: [
-        { questionNumber: 63, score: 2, correctAnswer: null },
-        { questionNumber: 64, score: 2, correctAnswer: null },
-        { questionNumber: 65, score: 2, correctAnswer: null },
-        { questionNumber: 66, score: 2, correctAnswer: null },
-        { questionNumber: 67, score: 2, correctAnswer: null },
-        { questionNumber: 68, score: 2, correctAnswer: null },
-        { questionNumber: 69, score: 2, correctAnswer: null },
-        { questionNumber: 70, score: 2, correctAnswer: null },
-        { questionNumber: 71, score: 2, correctAnswer: null },
-        { questionNumber: 72, score: 2, correctAnswer: null },
-      ],
-      sectionTotal: { score: 20 },
-    },
-  ],
-};
-
-// 2016年 数学II・数学B テストデータ
-const mockTestData_2B_2016: TestData = {
-  subject: "Math2B",
-  year: 2016,
-  maxScore: 100,
-  testStructure: [
-    {
-      section: 1,
-      questions: [
-        { questionNumber: 1, score: 2, correctAnswer: null },
-        { questionNumber: 2, score: 2, correctAnswer: null },
-        { questionNumber: 3, score: 1, correctAnswer: null },
-        { questionNumber: 4, score: 1, correctAnswer: null },
-        { questionNumber: 5, score: 1, correctAnswer: null },
-        { questionNumber: 6, score: 1, correctAnswer: null },
-        { questionNumber: 7, score: 2, correctAnswer: null },
-        { questionNumber: 8, score: 2, correctAnswer: null },
-        { questionNumber: 9, score: 2, correctAnswer: null },
-        { questionNumber: 10, score: 1, correctAnswer: null },
-        { questionNumber: 11, score: 3, correctAnswer: null },
-        { questionNumber: 12, score: 2, correctAnswer: null },
-        { questionNumber: 13, score: 2, correctAnswer: null },
-        { questionNumber: 14, score: 2, correctAnswer: null },
-        { questionNumber: 15, score: 2, correctAnswer: null },
-        { questionNumber: 16, score: 1, correctAnswer: null },
-        { questionNumber: 17, score: 1, correctAnswer: null },
-        { questionNumber: 18, score: 2, correctAnswer: null },
-      ],
-      sectionTotal: { score: 30 },
-    },
-    {
-      section: 2,
-      questions: [
-        { questionNumber: 19, score: 2, correctAnswer: null },
-        { questionNumber: 20, score: 3, correctAnswer: null },
-        { questionNumber: 21, score: 3, correctAnswer: null },
-        { questionNumber: 22, score: 2, correctAnswer: null },
-        { questionNumber: 23, score: 3, correctAnswer: null },
-        { questionNumber: 24, score: 2, correctAnswer: null },
-        { questionNumber: 25, score: 1, correctAnswer: null },
-        { questionNumber: 26, score: 2, correctAnswer: null },
-        { questionNumber: 27, score: 2, correctAnswer: null },
-        { questionNumber: 28, score: 2, correctAnswer: null },
-        { questionNumber: 29, score: 3, correctAnswer: null },
-        { questionNumber: 30, score: 3, correctAnswer: null },
-        { questionNumber: 31, score: 2, correctAnswer: null },
-      ],
-      sectionTotal: { score: 30 },
-    },
-    {
-      section: 3,
-      questions: [
-        { questionNumber: 32, score: 2, correctAnswer: null },
-        { questionNumber: 33, score: 2, correctAnswer: null },
-        { questionNumber: 34, score: 2, correctAnswer: null },
-        { questionNumber: 35, score: 2, correctAnswer: null },
-        { questionNumber: 36, score: 2, correctAnswer: null },
-        { questionNumber: 37, score: 2, correctAnswer: null },
-        { questionNumber: 38, score: 4, correctAnswer: null },
-        { questionNumber: 39, score: 4, correctAnswer: null },
-      ],
-      sectionTotal: { score: 20 },
-    },
-    {
-      section: 4,
-      questions: [
-        { questionNumber: 40, score: 1, correctAnswer: null },
-        { questionNumber: 41, score: 1, correctAnswer: null },
-        { questionNumber: 42, score: 2, correctAnswer: null },
-        { questionNumber: 43, score: 2, correctAnswer: null },
-        { questionNumber: 44, score: 1, correctAnswer: null },
-        { questionNumber: 45, score: 1, correctAnswer: null },
-        { questionNumber: 46, score: 1, correctAnswer: null },
-        { questionNumber: 47, score: 1, correctAnswer: null },
-        { questionNumber: 48, score: 1, correctAnswer: null },
-        { questionNumber: 49, score: 1, correctAnswer: null },
-        { questionNumber: 50, score: 2, correctAnswer: null },
-        { questionNumber: 51, score: 2, correctAnswer: null },
-        { questionNumber: 52, score: 2, correctAnswer: null },
-        { questionNumber: 53, score: 2, correctAnswer: null },
-      ],
-      sectionTotal: { score: 20 },
-    },
-    {
-      section: 5,
-      questions: [
-        { questionNumber: 54, score: 1, correctAnswer: null },
-        { questionNumber: 55, score: 1, correctAnswer: null },
-        { questionNumber: 56, score: 1, correctAnswer: null },
-        { questionNumber: 57, score: 1, correctAnswer: null },
-        { questionNumber: 58, score: 1, correctAnswer: null },
-        { questionNumber: 59, score: 1, correctAnswer: null },
-        { questionNumber: 60, score: 1, correctAnswer: null },
-        { questionNumber: 61, score: 1, correctAnswer: null },
-        { questionNumber: 62, score: 2, correctAnswer: null },
-        { questionNumber: 63, score: 1, correctAnswer: null },
-        { questionNumber: 64, score: 1, correctAnswer: null },
-        { questionNumber: 65, score: 1, correctAnswer: null },
-        { questionNumber: 66, score: 1, correctAnswer: null },
-        { questionNumber: 67, score: 2, correctAnswer: null },
-        { questionNumber: 68, score: 2, correctAnswer: null },
-        { questionNumber: 69, score: 2, correctAnswer: null },
-      ],
-      sectionTotal: { score: 20 },
-    },
-  ],
-};
-
-// 2017年 数学II・数学B テストデータ
-const mockTestData_2B_2017: TestData = {
-  subject: "Math2B",
-  year: 2017,
-  maxScore: 100,
-  testStructure: [
-    {
-      section: 1,
-      questions: [
-        { questionNumber: 1, score: 3, correctAnswer: null },
-        { questionNumber: 2, score: 2, correctAnswer: null },
-        { questionNumber: 3, score: 3, correctAnswer: null },
-        { questionNumber: 4, score: 3, correctAnswer: null },
-        { questionNumber: 5, score: 2, correctAnswer: null },
-        { questionNumber: 6, score: 2, correctAnswer: null },
-        { questionNumber: 7, score: 2, correctAnswer: null },
-        { questionNumber: 8, score: 2, correctAnswer: null },
-        { questionNumber: 9, score: 2, correctAnswer: null },
-        { questionNumber: 10, score: 3, correctAnswer: null },
-        { questionNumber: 11, score: 2, correctAnswer: null },
-        { questionNumber: 12, score: 2, correctAnswer: null },
-        { questionNumber: 13, score: 2, correctAnswer: null },
-      ],
-      sectionTotal: { score: 30 },
-    },
-    {
-      section: 2,
-      questions: [
-        { questionNumber: 14, score: 2, correctAnswer: null },
-        { questionNumber: 15, score: 1, correctAnswer: null },
-        { questionNumber: 16, score: 2, correctAnswer: null },
-        { questionNumber: 17, score: 1, correctAnswer: null },
-        { questionNumber: 18, score: 1, correctAnswer: null },
-        { questionNumber: 19, score: 1, correctAnswer: null },
-        { questionNumber: 20, score: 2, correctAnswer: null },
-        { questionNumber: 21, score: 2, correctAnswer: null },
-        { questionNumber: 22, score: 3, correctAnswer: null },
-        { questionNumber: 23, score: 3, correctAnswer: null },
-        { questionNumber: 24, score: 3, correctAnswer: null },
-        { questionNumber: 25, score: 3, correctAnswer: null },
-        { questionNumber: 26, score: 3, correctAnswer: null },
-        { questionNumber: 27, score: 3, correctAnswer: null },
-      ],
-      sectionTotal: { score: 30 },
-    },
-    {
-      section: 3,
-      questions: [
-        { questionNumber: 28, score: 2, correctAnswer: null },
-        { questionNumber: 29, score: 2, correctAnswer: null },
-        { questionNumber: 30, score: 2, correctAnswer: null },
-        { questionNumber: 31, score: 3, correctAnswer: null },
-        { questionNumber: 32, score: 2, correctAnswer: null },
-        { questionNumber: 33, score: 2, correctAnswer: null },
-        { questionNumber: 34, score: 2, correctAnswer: null },
-        { questionNumber: 35, score: 2, correctAnswer: null },
-        { questionNumber: 36, score: 2, correctAnswer: null },
-        { questionNumber: 37, score: 1, correctAnswer: null },
-      ],
-      sectionTotal: { score: 20 },
-    },
-    {
-      section: 4,
-      questions: [
-        { questionNumber: 38, score: 1, correctAnswer: null },
-        { questionNumber: 39, score: 1, correctAnswer: null },
-        { questionNumber: 40, score: 2, correctAnswer: null },
-        { questionNumber: 41, score: 2, correctAnswer: null },
-        { questionNumber: 42, score: 2, correctAnswer: null },
-        { questionNumber: 43, score: 2, correctAnswer: null },
-        { questionNumber: 44, score: 2, correctAnswer: null },
-        { questionNumber: 45, score: 2, correctAnswer: null },
-        { questionNumber: 46, score: 3, correctAnswer: null },
-        { questionNumber: 47, score: 3, correctAnswer: null },
-      ],
-      sectionTotal: { score: 20 },
-    },
-    {
-      section: 5,
-      questions: [
-        { questionNumber: 48, score: 3, correctAnswer: null },
-        { questionNumber: 49, score: 3, correctAnswer: null },
-        { questionNumber: 50, score: 3, correctAnswer: null },
-        { questionNumber: 51, score: 3, correctAnswer: null },
-        { questionNumber: 52, score: 2, correctAnswer: null },
-        { questionNumber: 53, score: 3, correctAnswer: null },
-        { questionNumber: 54, score: 2, correctAnswer: null },
-        { questionNumber: 55, score: 1, correctAnswer: null },
-      ],
-      sectionTotal: { score: 20 },
-    },
-  ],
-};
-
-// 2018年 数学II・数学B テストデータ
-const mockTestData_2B_2018: TestData = {
-  subject: "Math2B",
-  year: 2018,
-  maxScore: 100,
-  testStructure: [
-    {
-      section: 1,
-      questions: [
-        { questionNumber: 1, score: 1, correctAnswer: null },
-        { questionNumber: 2, score: 1, correctAnswer: null },
-        { questionNumber: 3, score: 2, correctAnswer: null },
-        { questionNumber: 4, score: 2, correctAnswer: null },
-        { questionNumber: 5, score: 2, correctAnswer: null },
-        { questionNumber: 6, score: 3, correctAnswer: null },
-        { questionNumber: 7, score: 3, correctAnswer: null },
-        { questionNumber: 8, score: 3, correctAnswer: null },
-        { questionNumber: 9, score: 2, correctAnswer: null },
-        { questionNumber: 10, score: 1, correctAnswer: null },
-        { questionNumber: 11, score: 1, correctAnswer: null },
-        { questionNumber: 12, score: 2, correctAnswer: null },
-        { questionNumber: 13, score: 3, correctAnswer: null },
-        { questionNumber: 14, score: 3, correctAnswer: null },
-      ],
-      sectionTotal: { score: 30 },
-    },
-    {
-      section: 2,
-      questions: [
-        { questionNumber: 15, score: 1, correctAnswer: null },
-        { questionNumber: 16, score: 2, correctAnswer: null },
-        { questionNumber: 17, score: 2, correctAnswer: null },
-        { questionNumber: 18, score: 4, correctAnswer: null },
-        { questionNumber: 19, score: 2, correctAnswer: null },
-        { questionNumber: 20, score: 3, correctAnswer: null },
-        { questionNumber: 21, score: 3, correctAnswer: null },
-        { questionNumber: 22, score: 1, correctAnswer: null },
-        { questionNumber: 23, score: 3, correctAnswer: null },
-        { questionNumber: 24, score: 4, correctAnswer: null },
-        { questionNumber: 25, score: 2, correctAnswer: null },
-        { questionNumber: 26, score: 3, correctAnswer: null },
-      ],
-      sectionTotal: { score: 30 },
-    },
-    {
-      section: 3,
-      questions: [
-        { questionNumber: 27, score: 2, correctAnswer: null },
-        { questionNumber: 28, score: 2, correctAnswer: null },
-        { questionNumber: 29, score: 2, correctAnswer: null },
-        { questionNumber: 30, score: 2, correctAnswer: null },
-        { questionNumber: 31, score: 2, correctAnswer: null },
-        { questionNumber: 32, score: 2, correctAnswer: null },
-        { questionNumber: 33, score: 2, correctAnswer: null },
-        { questionNumber: 34, score: 2, correctAnswer: null },
-        { questionNumber: 35, score: 1, correctAnswer: null },
-        { questionNumber: 36, score: 3, correctAnswer: null },
-      ],
-      sectionTotal: { score: 20 },
-    },
-    {
-      section: 4,
-      questions: [
-        { questionNumber: 37, score: 1, correctAnswer: null },
-        { questionNumber: 38, score: 1, correctAnswer: null },
-        { questionNumber: 39, score: 2, correctAnswer: null },
-        { questionNumber: 40, score: 2, correctAnswer: null },
-        { questionNumber: 41, score: 2, correctAnswer: null },
-        { questionNumber: 42, score: 4, correctAnswer: null },
-        { questionNumber: 43, score: 2, correctAnswer: null },
-        { questionNumber: 44, score: 2, correctAnswer: null },
-        { questionNumber: 45, score: 2, correctAnswer: null },
-        { questionNumber: 46, score: 3, correctAnswer: null },
-      ],
-      sectionTotal: { score: 20 },
-    },
-    {
-      section: 5,
-      questions: [
-        { questionNumber: 47, score: 2, correctAnswer: null },
-        { questionNumber: 48, score: 1, correctAnswer: null },
-        { questionNumber: 49, score: 1, correctAnswer: null },
-        { questionNumber: 50, score: 1, correctAnswer: null },
-        { questionNumber: 51, score: 1, correctAnswer: null },
-        { questionNumber: 52, score: 1, correctAnswer: null },
-        { questionNumber: 53, score: 2, correctAnswer: null },
-        { questionNumber: 54, score: 1, correctAnswer: null },
-        { questionNumber: 55, score: 1, correctAnswer: null },
-        { questionNumber: 56, score: 1, correctAnswer: null },
-        { questionNumber: 57, score: 1, correctAnswer: null },
-        { questionNumber: 58, score: 2, correctAnswer: null },
-        { questionNumber: 59, score: 1, correctAnswer: null },
-        { questionNumber: 60, score: 1, correctAnswer: null },
-        { questionNumber: 61, score: 1, correctAnswer: null },
-        { questionNumber: 62, score: 2, correctAnswer: null },
-      ],
-      sectionTotal: { score: 20 },
-    },
-  ],
-};
-
-// 2019年 数学II・数学B テストデータ
-const mockTestData_2B_2019: TestData = {
-  subject: "Math2B",
-  year: 2019,
-  maxScore: 100,
-  testStructure: [
-    {
-      section: 1,
-      questions: [
-        { questionNumber: 1, score: 1, correctAnswer: null },
-        { questionNumber: 2, score: 2, correctAnswer: null },
-        { questionNumber: 3, score: 2, correctAnswer: null },
-        { questionNumber: 4, score: 3, correctAnswer: null },
-        { questionNumber: 5, score: 3, correctAnswer: null },
-        { questionNumber: 6, score: 2, correctAnswer: null },
-        { questionNumber: 7, score: 2, correctAnswer: null },
-        { questionNumber: 8, score: 2, correctAnswer: null },
-        { questionNumber: 9, score: 2, correctAnswer: null },
-        { questionNumber: 10, score: 2, correctAnswer: null },
-        { questionNumber: 11, score: 2, correctAnswer: null },
-        { questionNumber: 12, score: 1, correctAnswer: null },
-        { questionNumber: 13, score: 1, correctAnswer: null },
-        { questionNumber: 14, score: 1, correctAnswer: null },
-        { questionNumber: 15, score: 2, correctAnswer: null },
-        { questionNumber: 16, score: 2, correctAnswer: null },
-      ],
-      sectionTotal: { score: 30 },
-    },
-    {
-      section: 2,
-      questions: [
-        { questionNumber: 17, score: 1, correctAnswer: null },
-        { questionNumber: 18, score: 1, correctAnswer: null },
-        { questionNumber: 19, score: 1, correctAnswer: null },
-        { questionNumber: 20, score: 2, correctAnswer: null },
-        { questionNumber: 21, score: 1, correctAnswer: null },
-        { questionNumber: 22, score: 2, correctAnswer: null },
-        { questionNumber: 23, score: 2, correctAnswer: null },
-        { questionNumber: 24, score: 2, correctAnswer: null },
-        { questionNumber: 25, score: 2, correctAnswer: null },
-        { questionNumber: 26, score: 3, correctAnswer: null },
-        { questionNumber: 27, score: 2, correctAnswer: null },
-        { questionNumber: 28, score: 1, correctAnswer: null },
-        { questionNumber: 29, score: 2, correctAnswer: null },
-        { questionNumber: 30, score: 3, correctAnswer: null },
-        { questionNumber: 31, score: 3, correctAnswer: null },
-        { questionNumber: 32, score: 3, correctAnswer: null },
-      ],
-      sectionTotal: { score: 30 },
-    },
-    {
-      section: 3,
-      questions: [
-        { questionNumber: 33, score: 2, correctAnswer: null },
-        { questionNumber: 34, score: 2, correctAnswer: null },
-        { questionNumber: 35, score: 2, correctAnswer: null },
-        { questionNumber: 36, score: 3, correctAnswer: null },
-        { questionNumber: 37, score: 1, correctAnswer: null },
-        { questionNumber: 38, score: 3, correctAnswer: null },
-        { questionNumber: 39, score: 2, correctAnswer: null },
-        { questionNumber: 40, score: 2, correctAnswer: null },
-        { questionNumber: 41, score: 3, correctAnswer: null },
-      ],
-      sectionTotal: { score: 20 },
-    },
-    {
-      section: 4,
-      questions: [
-        { questionNumber: 42, score: 1, correctAnswer: null },
-        { questionNumber: 43, score: 1, correctAnswer: null },
-        { questionNumber: 44, score: 1, correctAnswer: null },
-        { questionNumber: 45, score: 1, correctAnswer: null },
-        { questionNumber: 46, score: 1, correctAnswer: null },
-        { questionNumber: 47, score: 1, correctAnswer: null },
-        { questionNumber: 48, score: 1, correctAnswer: null },
-        { questionNumber: 49, score: 1, correctAnswer: null },
-        { questionNumber: 50, score: 1, correctAnswer: null },
-        { questionNumber: 51, score: 2, correctAnswer: null },
-        { questionNumber: 52, score: 1, correctAnswer: null },
-        { questionNumber: 53, score: 2, correctAnswer: null },
-        { questionNumber: 54, score: 2, correctAnswer: null },
-        { questionNumber: 55, score: 2, correctAnswer: null },
-        { questionNumber: 56, score: 2, correctAnswer: null },
-      ],
-      sectionTotal: { score: 20 },
-    },
-    {
-      section: 5,
-      questions: [
-        { questionNumber: 57, score: 2, correctAnswer: null },
-        { questionNumber: 58, score: 2, correctAnswer: null },
-        { questionNumber: 59, score: 2, correctAnswer: null },
-        { questionNumber: 60, score: 2, correctAnswer: null },
-        { questionNumber: 61, score: 2, correctAnswer: null },
-        { questionNumber: 62, score: 2, correctAnswer: null },
-        { questionNumber: 63, score: 2, correctAnswer: null },
-        { questionNumber: 64, score: 2, correctAnswer: null },
-        { questionNumber: 65, score: 2, correctAnswer: null },
-        { questionNumber: 66, score: 2, correctAnswer: null },
-      ],
-      sectionTotal: { score: 20 },
-    },
-  ],
-};
-
-// 2020年 数学II・数学B テストデータの続き
-const mockTestData_2B_2020: TestData = {
-  subject: "Math2B",
-  year: 2020,
-  maxScore: 100,
-  testStructure: [
-    {
-      section: 1,
-      questions: [
-        { questionNumber: 1, score: 2, correctAnswer: null },
-        { questionNumber: 2, score: 2, correctAnswer: null },
-        { questionNumber: 3, score: 3, correctAnswer: null },
-        { questionNumber: 4, score: 2, correctAnswer: null },
-        { questionNumber: 5, score: 2, correctAnswer: null },
-        { questionNumber: 6, score: 2, correctAnswer: null },
-        { questionNumber: 7, score: 2, correctAnswer: null },
-        { questionNumber: 8, score: 3, correctAnswer: null },
-        { questionNumber: 9, score: 2, correctAnswer: null },
-        { questionNumber: 10, score: 2, correctAnswer: null },
-        { questionNumber: 11, score: 2, correctAnswer: null },
-        { questionNumber: 12, score: 2, correctAnswer: null },
-        { questionNumber: 13, score: 2, correctAnswer: null },
-        { questionNumber: 14, score: 2, correctAnswer: null },
-      ],
-      sectionTotal: { score: 30 },
-    },
-    {
-      section: 2,
-      questions: [
-        { questionNumber: 15, score: 2, correctAnswer: null },
-        { questionNumber: 16, score: 2, correctAnswer: null },
-        { questionNumber: 17, score: 2, correctAnswer: null },
-        { questionNumber: 18, score: 2, correctAnswer: null },
-        { questionNumber: 19, score: 3, correctAnswer: null },
-        { questionNumber: 20, score: 2, correctAnswer: null },
-        { questionNumber: 21, score: 2, correctAnswer: null },
-        { questionNumber: 22, score: 3, correctAnswer: null },
-        { questionNumber: 23, score: 3, correctAnswer: null },
-        { questionNumber: 24, score: 3, correctAnswer: null },
-        { questionNumber: 25, score: 3, correctAnswer: null },
-        { questionNumber: 26, score: 3, correctAnswer: null },
-      ],
-      sectionTotal: { score: 30 },
-    },
-    {
-      section: 3,
-      questions: [
-        { questionNumber: 27, score: 2, correctAnswer: null },
-        { questionNumber: 28, score: 1, correctAnswer: null },
-        { questionNumber: 29, score: 2, correctAnswer: null },
-        { questionNumber: 30, score: 1, correctAnswer: null },
-        { questionNumber: 31, score: 1, correctAnswer: null },
-        { questionNumber: 32, score: 2, correctAnswer: null },
-        { questionNumber: 33, score: 2, correctAnswer: null },
-        { questionNumber: 34, score: 2, correctAnswer: null },
-        { questionNumber: 35, score: 2, correctAnswer: null },
-        { questionNumber: 36, score: 2, correctAnswer: null },
-        { questionNumber: 37, score: 1, correctAnswer: null },
-        { questionNumber: 38, score: 2, correctAnswer: null },
-      ],
-      sectionTotal: { score: 20 },
-    },
-    {
-      section: 4,
-      questions: [
-        { questionNumber: 39, score: 2, correctAnswer: null },
-        { questionNumber: 40, score: 2, correctAnswer: null },
-        { questionNumber: 41, score: 2, correctAnswer: null },
-        { questionNumber: 42, score: 1, correctAnswer: null },
-        { questionNumber: 43, score: 1, correctAnswer: null },
-        { questionNumber: 44, score: 1, correctAnswer: null },
-        { questionNumber: 45, score: 2, correctAnswer: null },
-        { questionNumber: 46, score: 1, correctAnswer: null },
-        { questionNumber: 47, score: 1, correctAnswer: null },
-        { questionNumber: 48, score: 2, correctAnswer: null },
-        { questionNumber: 49, score: 2, correctAnswer: null },
-        { questionNumber: 50, score: 1, correctAnswer: null },
-        { questionNumber: 51, score: 1, correctAnswer: null },
-        { questionNumber: 52, score: 1, correctAnswer: null },
-      ],
-      sectionTotal: { score: 20 },
-    },
-    {
-      section: 5,
-      questions: [
-        { questionNumber: 53, score: 2, correctAnswer: null },
-        { questionNumber: 54, score: 2, correctAnswer: null },
-        { questionNumber: 55, score: 2, correctAnswer: null },
-        { questionNumber: 56, score: 2, correctAnswer: null },
-        { questionNumber: 57, score: 2, correctAnswer: null },
-        { questionNumber: 58, score: 2, correctAnswer: null },
-        { questionNumber: 59, score: 2, correctAnswer: null },
-        { questionNumber: 60, score: 2, correctAnswer: null },
-        { questionNumber: 61, score: 2, correctAnswer: null },
-        { questionNumber: 62, score: 2, correctAnswer: null },
-      ],
-      sectionTotal: { score: 20 },
-    },
-  ],
-};
-
-const profileData = {
-  userName: "数1Aは32点だけど千葉大に受かりたい人",
-  targetUniversity: [
-    "千葉大学-理学部-地球科学科",
-    "明治大学-農学部-農学科",
-    "東京農業大学-農学部-農学科",
-  ],
-  memo: "サッカー部の活動時間が長く、勉強時間が1日2時間程度しか取れていないため、内職をして時間を確保したり、過去問をもとに本当に必要なものを優先したりと工夫して勉強しています。",
-};
-
-const cardDatas = [
-  {
-    subject: "Reading",
-    finalScoreTarget: 84,
-    finalScoreLowest: 70,
-    memo: "英語は1月から英語長文ポラリス1の文を例文としてスラスラ言えるレベルをテーマに30分/日音読をしたことで伸びた気がします。",
-    testResults: [
-      {
-        id: 8,
-        date: "2024/08/01",
-        year: "2018",
-        targetScore: 80,
-        studentScore: 75,
-        memo: "この日は調子が悪かったのでしょうがなかった。",
-      },
-      {
-        id: 7,
-        date: "2024/07/01",
-        year: "2017",
-        targetScore: 70,
-        studentScore: 65,
-        memo: "この日は調子が悪かったのでしょうがなかった。",
-      },
-      {
-        id: 6,
-        date: "2024/06/01",
-        year: "2016",
-        targetScore: 85,
-        studentScore: 80,
-        memo: "この日は調子が悪かったのでしょうがなかった。",
-      },
-      {
-        id: 5,
-        date: "2024/05/01",
-        year: "2015",
-        targetScore: 90,
-        studentScore: 85,
-        memo: "この日は調子が悪かったのでしょうがなかった。",
-      },
-      {
-        id: 4,
-        date: "2024/04/01",
-        year: "2014",
-        targetScore: 80,
-        studentScore: 75,
-        memo: "この日は調子が悪かったのでしょうがなかった。",
-      },
-      {
-        id: 3,
-        date: "2024/03/01",
-        year: "2013",
-        targetScore: 60,
-        studentScore: 55,
-        memo: "この日は調子が悪かったのでしょうがなかった。",
-      },
-      {
-        id: 2,
-        date: "2024/02/01",
-        year: "2012",
-        targetScore: 40,
-        studentScore: 45,
-        memo: "この日は調子が悪かったのでしょうがなかった。",
-      },
-      {
-        id: 1,
-        date: "2024/01/01",
-        year: "2011",
-        targetScore: 10,
-        studentScore: 15,
-        memo: "この日は調子が悪かったのでしょうがなかった。",
-      },
-    ],
-  },
-  {
-    subject: "Math1A",
-    finalScoreTarget: 75,
-    finalScoreLowest: 60,
-    memo: "数1Aは基礎問を2周回したら70点台に乗りました。とてもおすすめなのでぜひやってみてください。",
-    testResults: [
-      {
-        id: 8,
-        date: "2024/08/01",
-        year: "2018",
-        targetScore: 80,
-        studentScore: 75,
-        memo: "この日は調子が悪かったのでしょうがなかった。",
-      },
-      {
-        id: 7,
-        date: "2024/07/01",
-        year: "2017",
-        targetScore: 70,
-        studentScore: 65,
-        memo: "この日は調子が悪かったのでしょうがなかった。",
-      },
-      {
-        id: 6,
-        date: "2024/06/01",
-        year: "2016",
-        targetScore: 85,
-        studentScore: 80,
-        memo: "この日は調子が悪かったのでしょうがなかった。",
-      },
-      {
-        id: 5,
-        date: "2024/05/01",
-        year: "2015",
-        targetScore: 90,
-        studentScore: 85,
-        memo: "この日は調子が悪かったのでしょうがなかった。",
-      },
-      {
-        id: 4,
-        date: "2024/04/01",
-        year: "2014",
-        targetScore: 80,
-        studentScore: 75,
-        memo: "この日は調子が悪かったのでしょうがなかった。",
-      },
-      {
-        id: 3,
-        date: "2024/03/01",
-        year: "2013",
-        targetScore: 60,
-        studentScore: 55,
-        memo: "この日は調子が悪かったのでしょうがなかった。",
-      },
-      {
-        id: 2,
-        date: "2024/02/01",
-        year: "2012",
-        targetScore: 40,
-        studentScore: 45,
-        memo: "この日は調子が悪かったのでしょうがなかった。",
-      },
-      {
-        id: 1,
-        date: "2024/01/01",
-        year: "2011",
-        targetScore: 10,
-        studentScore: 15,
-        memo: "この日は調子が悪かったのでしょうがなかった。",
-      },
-    ],
-  },
-  {
-    subject: "Math2B",
-    finalScoreTarget: 75,
-    finalScoreLowest: 60,
-    memo: "数2Bは基礎問を3周回したら70点台に乗りました。たくさん演習を積むことが大事だと思います。",
-    testResults: [
-      {
-        id: 8,
-        date: "2024/08/01",
-        year: "2018",
-        targetScore: 80,
-        studentScore: 75,
-        memo: "この日は調子が悪かったのでしょうがなかった。",
-      },
-      {
-        id: 7,
-        date: "2024/07/01",
-        year: "2017",
-        targetScore: 70,
-        studentScore: 65,
-        memo: "この日は調子が悪かったのでしょうがなかった。",
-      },
-      {
-        id: 6,
-        date: "2024/06/01",
-        year: "2016",
-        targetScore: 85,
-        studentScore: 80,
-        memo: "この日は調子が悪かったのでしょうがなかった。",
-      },
-      {
-        id: 5,
-        date: "2024/05/01",
-        year: "2015",
-        targetScore: 90,
-        studentScore: 85,
-        memo: "この日は調子が悪かったのでしょうがなかった。",
-      },
-      {
-        id: 4,
-        date: "2024/04/01",
-        year: "2014",
-        targetScore: 80,
-        studentScore: 75,
-        memo: "この日は調子が悪かったのでしょうがなかった。",
-      },
-      {
-        id: 3,
-        date: "2024/03/01",
-        year: "2013",
-        targetScore: 60,
-        studentScore: 55,
-        memo: "この日は調子が悪かったのでしょうがなかった。",
-      },
-      {
-        id: 2,
-        date: "2024/02/01",
-        year: "2012",
-        targetScore: 40,
-        studentScore: 45,
-        memo: "この日は調子が悪かったのでしょうがなかった。",
-      },
-      {
-        id: 1,
-        date: "2024/01/01",
-        year: "2011",
-        targetScore: 10,
-        studentScore: 15,
-        memo: "この日は調子が悪かったのでしょうがなかった。",
-      },
-    ],
-  },
-  {
-    subject: "Chemistry",
-    finalScoreTarget: 75,
-    finalScoreLowest: 65,
-    memo: "共通テストの化学は本当に難しいので目標の点数は低めに設定しました。",
-    testResults: [
-      {
-        id: 8,
-        date: "2024/08/01",
-        year: "2018",
-        targetScore: 80,
-        studentScore: 75,
-        memo: "この日は調子が悪かったのでしょうがなかった。",
-      },
-      {
-        id: 7,
-        date: "2024/07/01",
-        year: "2017",
-        targetScore: 70,
-        studentScore: 65,
-        memo: "この日は調子が悪かったのでしょうがなかった。",
-      },
-      {
-        id: 6,
-        date: "2024/06/01",
-        year: "2016",
-        targetScore: 85,
-        studentScore: 80,
-        memo: "この日は調子が悪かったのでしょうがなかった。",
-      },
-      {
-        id: 5,
-        date: "2024/05/01",
-        year: "2015",
-        targetScore: 90,
-        studentScore: 85,
-        memo: "この日は調子が悪かったのでしょうがなかった。",
-      },
-      {
-        id: 4,
-        date: "2024/04/01",
-        year: "2014",
-        targetScore: 80,
-        studentScore: 75,
-        memo: "この日は調子が悪かったのでしょうがなかった。",
-      },
-      {
-        id: 3,
-        date: "2024/03/01",
-        year: "2013",
-        targetScore: 60,
-        studentScore: 55,
-        memo: "この日は調子が悪かったのでしょうがなかった。",
-      },
-      {
-        id: 2,
-        date: "2024/02/01",
-        year: "2012",
-        targetScore: 40,
-        studentScore: 45,
-        memo: "この日は調子が悪かったのでしょうがなかった。",
-      },
-      {
-        id: 1,
-        date: "2024/01/01",
-        year: "2011",
-        targetScore: 10,
-        studentScore: 15,
-        memo: "この日は調子が悪かったのでしょうがなかった。",
-      },
-    ],
-  },
-];
-
-const mockTestDatas = [
-  mockTestData_R_2015,
-  mockTestData_R_2016,
-  mockTestData_R_2017,
-  mockTestData_R_2018,
-  mockTestData_R_2019,
-  mockTestData_R_2020,
-  mockTestData_1A_2015,
-  mockTestData_1A_2016,
-  mockTestData_1A_2017,
-  mockTestData_1A_2018,
-  mockTestData_1A_2019,
-  mockTestData_1A_2020,
-  mockTestData_2B_2015,
-  mockTestData_2B_2016,
-  mockTestData_2B_2017,
-  mockTestData_2B_2018,
-  mockTestData_2B_2019,
-  mockTestData_2B_2020,
-];
+// src/core/profile/profileService.ts
+import { testDatas } from "./repositories/data";
+import { ProfileRepository } from "./repositories/infra";
+import { ProfileData, CardData, AnsweredData, TestData } from "./type";
 
 class ProfileService {
+  private repository: ProfileRepository;
+
+  constructor() {
+    this.repository = new ProfileRepository();
+  }
+
+  // プロフィール情報の取得
   async fetchProfileData(): Promise<ProfileData> {
-    return new Promise((resolve) => {
-      setTimeout(() => resolve(profileData), 100);
-    });
+    return this.repository.fetchProfileData();
   }
+
+  // 科目カード情報の取得
   async fetchCardDatas(): Promise<CardData[]> {
-    return new Promise((resolve) => {
-      setTimeout(() => resolve(cardDatas), 100);
-    });
-  }
-  // 生徒データを取得するメソッド
-  async fetchStudentData(subject: string, year: number): Promise<StudentData> {
-    return new Promise((resolve) => {
-      setTimeout(() => resolve(mockStudentData_R_2018), 100);
-    });
+    return this.repository.fetchCardDatas();
   }
 
-  // フレンドデータを取得するメソッド
-  async fetchFriendsData(subject: string, year: number): Promise<Friend[]> {
-    return new Promise((resolve) => {
-      setTimeout(() => resolve(mockFriendsData_R_2018), 100);
-    });
-  }
-
-  // テスト構造を取得するメソッド
+  // テスト構造情報の取得
   async fetchTestStructure(subject: string, year: number): Promise<TestData> {
-    return new Promise((resolve, reject) => {
-      // subjectとyearを使ってモックデータを検索
-      const matchedTest = mockTestDatas.find(
-        (test) => test.subject === subject && test.year === year
-      );
-
-      if (matchedTest) {
-        // 一致するテストデータがある場合、それを返す
-        setTimeout(() => resolve(matchedTest), 100);
-      } else {
-        // 一致するテストデータがない場合はエラーを返す
-        setTimeout(
-          () => reject(new Error(`Test data for ${subject} ${year} not found`)),
-          100
-        );
-      }
-    });
+    const testData = testDatas.find(
+      (testData) => testData.subject === subject && testData.year === year
+    );
+    if (!testData) {
+      throw new Error(`Test data not found for subject: ${subject}, year: ${year}`);
+    }
+    return testData;
   }
 
-  // レーダーチャートデータ（生徒）を取得するメソッド
-  async fetchStudentRaderChartData(subject: string, year: number) {
-    // これは実際には必要ないかもしれないが、将来的な拡張のために残しておく
-    const studentData = await this.fetchStudentData(subject, year);
-    return {
-      sectionPercentages: studentData.sectionPercentages,
-      targetSectionPercentages: studentData.targetSectionPercentages,
-    };
+  // 生徒のテスト結果データの取得
+  async fetchStudentData(subject: string, year: number): Promise<AnsweredData> {
+    return this.repository.fetchStudentData(subject, year);
   }
 
-  // レーダーチャートデータ（フレンド）を取得するメソッド
-  async fetchFriendRaderChartData(subject: string, year: number) {
-    // これは実際には必要ないかもしれないが、将来的な拡張のために残しておく
-    const friendsData = await this.fetchFriendsData(subject, year);
-    return friendsData.map((friend) => ({
-      id: friend.id,
-      name: friend.name,
-      sectionPercentages: friend.sectionPercentages,
-      targetSectionPercentages: friend.targetSectionPercentages,
-    }));
+  // フレンドのテスト結果データの取得
+  async fetchFriendsData(
+    subject: string,
+    year: number
+  ): Promise<AnsweredData[]> {
+    return this.repository.fetchFriendsData(subject, year);
   }
 }
 
-// シングルトンインスタンスをエクスポート
+// シングルトンとしてエクスポート
 export const profileService = new ProfileService();
