+++ tsconfig.scripts.json
@@ -1,0 +1,10 @@
+{
+    "compilerOptions": {
+      "target": "es2017",
+      "module": "commonjs",
+      "esModuleInterop": true,
+      "outDir": "dist",
+      "strict": true
+    },
+    "include": ["scripts/**/*"]
+  }
